+++ libs/deps/mas/merch-card.js
@@ -1,6 +1,6 @@
-import{html as o,LitElement as K,nothing as Q}from"/libs/deps/lit-all.min.js";import{LitElement as F,html as T,css as V}from"/libs/deps/lit-all.min.js";var m=class extends F{static properties={size:{type:String,attribute:!0},src:{type:String,attribute:!0},alt:{type:String,attribute:!0},href:{type:String,attribute:!0}};constructor(){super(),this.size="m",this.alt=""}render(){let{href:e}=this;return e?T`<a href="${e}">
+import{html as Pt,LitElement as Nt,nothing as Dt}from"/libs/deps/lit-all.min.js";import{LitElement as yt,html as F,css as wt}from"/libs/deps/lit-all.min.js";var a=class extends yt{static properties={size:{type:String,attribute:!0},src:{type:String,attribute:!0},alt:{type:String,attribute:!0},href:{type:String,attribute:!0}};constructor(){super(),this.size="m",this.alt=""}render(){let{href:t}=this;return t?F`<a href="${t}">
                   <img src="${this.src}" alt="${this.alt}" loading="lazy" />
-              </a>`:T` <img src="${this.src}" alt="${this.alt}" loading="lazy" />`}static styles=V`
+              </a>`:F` <img src="${this.src}" alt="${this.alt}" loading="lazy" />`}static styles=wt`
         :host {
             --img-width: 32px;
             --img-height: 32px;
@@ -23,7 +23,7 @@ import{html as o,LitElement as K,nothing as Q}from"/libs/deps/lit-all.min.js";im
             width: var(--img-width);
             height: var(--img-height);
         }
-    `};customElements.define("merch-icon",m);var U=localStorage.getItem("masAccessToken"),A={Authorization:`Bearer ${U}`,pragma:"no-cache","cache-control":"no-cache"};async function G({path:c,query:e}){let t={};c&&(t.path=c),e&&(t.fullText={text:encodeURIComponent(e),queryMode:"EXACT_WORDS"});let r=new URLSearchParams({query:JSON.stringify({filter:t})}).toString();return fetch(`${this.cfSearchUrl}?${r}`,{headers:A}).then(n=>n.json()).then(({items:n})=>n)}async function W(c){return fetch(`${this.cfFragmentsUrl}?path=${c}`,{headers:A}).then(e=>e.json()).then(({items:[e]})=>e)}var w=class{sites={cf:{fragments:{search:G.bind(this),getCfByPath:W.bind(this)}}};constructor(e){let r=`${`https://${e}.adobeaemcloud.com`}/adobe/sites`;this.cfFragmentsUrl=`${r}/cf/fragments`,this.cfSearchUrl=`${this.cfFragmentsUrl}/search`}};var _="merch-offer-select:ready",M="merch-card:ready",R="merch-card:action-menu-toggle";var C="merch-storage:change",k="merch-quantity-selector:change";function L(c){let e=[];function t(r){r.nodeType===Node.TEXT_NODE?e.push(r):r.childNodes.forEach(t)}return t(c),e}function x(c,e={},t){let r=document.createElement(c);r.innerHTML=t;for(let[n,a]of Object.entries(e))r.setAttribute(n,a);return r}function O(){return window.matchMedia("(max-width: 767px)").matches}function N(){return window.matchMedia("(max-width: 1024px)").matches}var H="aem-bucket",B={catalog:{name:"catalog",title:{tag:"h3",slot:"heading-xs"},prices:{tag:"h3",slot:"heading-xs"},description:{tag:"div",slot:"body-xs"},ctas:{size:"l"}},ah:{name:"ah",title:{tag:"h3",slot:"heading-xxs"},prices:{tag:"h3",slot:"heading-xs"},description:{tag:"div",slot:"body-xxs"},ctas:{size:"s"}},"ccd-action":{name:"ccd-action",title:{tag:"h3",slot:"heading-xs"},prices:{tag:"h3",slot:"heading-xs"},description:{tag:"div",slot:"body-xs"},ctas:{size:"l"}}};async function Y(c,e){let t=c.fields.reduce((a,{name:i,multiple:u,values:$})=>(a[i]=u?$:$[0],a),{}),{type:r="catalog"}=t,n=B[r]||B.catalog;if(e.variant=r,e.setAttribute("variant",r),t.icon?.forEach(a=>{let i=x("merch-icon",{slot:"icons",src:a,alt:"",href:"",size:"l"});e.append(i)}),t.title&&e.append(x(n.title.tag,{slot:n.title.slot},t.title)),t.prices){let a=t.prices,i=x(n.prices.tag,{slot:n.prices.slot},a);e.append(i)}if(e.append(x("p",{slot:"body-xxs",id:"individuals1"},"Desktop")),t.description){let a=x(n.description.tag,{slot:n.description.slot},t.description);e.append(a)}if(t.ctas){let a=t.ctas,i=x("div",{slot:"footer"},a);e.append(i)}}var z=class{#e=new Map;clear(){this.#e.clear()}add(...e){e.forEach(t=>{let{path:r}=t;r&&this.#e.set(r,t)})}has(e){return this.#e.has(e)}get(e){return this.#e.get(e)}remove(e){this.#e.delete(e)}},P=new z,l=class extends HTMLElement{#e;cache=P;path;static get observedAttributes(){return["source","path"]}attributeChangedCallback(e,t,r){this[e]=r}connectedCallback(){let e=this.getAttribute(H)??document.querySelector("mas-studio")?.getAttribute(H);this.#e=new w(e),this.fetchData()}refresh(){this.cache.remove(this.path),this.fetchData()}async fetchData(){let e=P.get(this.path);if(e||(e=await this.#e.sites.cf.fragments.getCfByPath(this.path)),e){Y(e,this.parentElement),this.render();return}this.render()}async render(){this.isConnected&&this.parentElement.tagName==="MERCH-CARD"&&await Promise.all([...this.parentElement.querySelectorAll('[is="inline-price"],[is="checkout-link"]')].map(e=>e.onceSettled()))}};customElements.define("merch-datasource",l);import{css as S,unsafeCSS as y}from"/libs/deps/lit-all.min.js";var v="(max-width: 767px)",b="(max-width: 1199px)",h="(min-width: 768px)",s="(min-width: 1200px)",d="(min-width: 1600px)";var D=S`
+    `};customElements.define("merch-icon",a);var Et=localStorage.getItem("masAccessToken"),K={Authorization:`Bearer ${Et}`,pragma:"no-cache","cache-control":"no-cache"};async function St({path:o,query:t}){let e={};o&&(e.path=o),t&&(e.fullText={text:encodeURIComponent(t),queryMode:"EXACT_WORDS"});let r=new URLSearchParams({query:JSON.stringify({filter:e})}).toString();return fetch(`${this.cfSearchUrl}?${r}`,{headers:K}).then(n=>n.json()).then(({items:n})=>n)}async function Ct(o){return fetch(`${this.cfFragmentsUrl}?path=${o}`,{headers:K}).then(t=>t.json()).then(({items:[t]})=>t)}var L=class{sites={cf:{fragments:{search:St.bind(this),getCfByPath:Ct.bind(this)}}};constructor(t){let r=`${`https://${t}.adobeaemcloud.com`}/adobe/sites`;this.cfFragmentsUrl=`${r}/cf/fragments`,this.cfSearchUrl=`${this.cfFragmentsUrl}/search`}};var W="merch-offer-select:ready",Y="merch-card:ready",Q="merch-card:action-menu-toggle";var P="merch-storage:change",N="merch-quantity-selector:change";function Z(o){let t=[];function e(r){r.nodeType===Node.TEXT_NODE?t.push(r):r.childNodes.forEach(e)}return e(o),t}function x(o,t={},e){let r=document.createElement(o);r.innerHTML=e;for(let[n,d]of Object.entries(t))r.setAttribute(n,d);return r}function J(){return window.matchMedia("(max-width: 767px)").matches}function X(){return window.matchMedia("(max-width: 1024px)").matches}var tt="aem-bucket",et={catalog:{name:"catalog",title:{tag:"h3",slot:"heading-xs"},prices:{tag:"h3",slot:"heading-xs"},description:{tag:"div",slot:"body-xs"},ctas:{size:"l"}},ah:{name:"ah",title:{tag:"h3",slot:"heading-xxs"},prices:{tag:"h3",slot:"heading-xs"},description:{tag:"div",slot:"body-xxs"},ctas:{size:"s"}},"ccd-action":{name:"ccd-action",title:{tag:"h3",slot:"heading-xs"},prices:{tag:"h3",slot:"heading-xs"},description:{tag:"div",slot:"body-xs"},ctas:{size:"l"}}};async function kt(o,t){let e=Array.isArray(o.fields)?o.fields.reduce((d,{name:u,multiple:v,values:I})=>(d[u]=v?I:I[0],d),{}):{},{type:r="catalog"}=e,n=et[r]||et.catalog;if(t.variant=r,t.setAttribute("variant",r),e.icon?.forEach(d=>{let u=x("merch-icon",{slot:"icons",src:d,alt:"",href:"",size:"l"});t.append(u)}),e.title&&t.append(x(n.title.tag,{slot:n.title.slot},e.title)),e.prices){let d=e.prices,u=x(n.prices.tag,{slot:n.prices.slot},d);t.append(u)}if(t.append(x("p",{slot:"body-xxs",id:"individuals1"},"Desktop")),e.description){let d=x(n.description.tag,{slot:n.description.slot},e.description);t.append(d)}if(e.ctas){let d=e.ctas,u=x("div",{slot:"footer"},d);t.append(u)}}var D=class{#t=new Map;clear(){this.#t.clear()}add(...t){t.forEach(e=>{let{path:r}=e;r&&this.#t.set(r,e)})}has(t){return this.#t.has(t)}get(t){return this.#t.get(t)}remove(t){this.#t.delete(t)}},rt=new D,c=class extends HTMLElement{#t;cache=rt;path;static get observedAttributes(){return["source","path"]}attributeChangedCallback(t,e,r){this[t]=r}connectedCallback(){let t=this.getAttribute(tt)??document.querySelector("mas-studio")?.getAttribute(tt);this.#t=new L(t),this.fetchData()}refresh(){this.cache.remove(this.path),this.fetchData()}async fetchData(){let t=rt.get(this.path);if(t||(t=await this.#t.sites.cf.fragments.getCfByPath(this.path)),t){kt(t,this.parentElement),this.render();return}this.render()}async render(){this.isConnected&&this.parentElement.tagName==="MERCH-CARD"&&await Promise.all([...this.parentElement.querySelectorAll('[is="inline-price"],[is="checkout-link"]')].map(t=>t.onceSettled()))}};customElements.define("merch-datasource",c);import{css as B,unsafeCSS as H}from"/libs/deps/lit-all.min.js";var f="(max-width: 767px)",_="(max-width: 1199px)",l="(min-width: 768px)",m="(min-width: 1200px)",g="(min-width: 1600px)";var ot=B`
     :host {
         position: relative;
         display: flex;
@@ -40,26 +40,6 @@ import{html as o,LitElement as K,nothing as Q}from"/libs/deps/lit-all.min.js";im
 
     :host(.placeholder) {
         visibility: hidden;
-    }
-
-    :host([variant='special-offers']) {
-        min-height: 439px;
-    }
-
-    :host([variant='catalog']) {
-        min-height: 330px;
-    }
-
-    :host([variant='plans']) {
-        min-height: 348px;
-    }
-
-    :host([variant='segment']) {
-        min-height: 214px;
-    }
-
-    :host([variant='ccd-action']:not([size])) {
-        width: var(--consonant-merch-card-ccd-action-width);
     }
 
     :host([aria-selected]) {
@@ -81,10 +61,6 @@ import{html as o,LitElement as K,nothing as Q}from"/libs/deps/lit-all.min.js";im
     .action-menu.always-visible {
         visibility: visible;
         background-image: var(--ellipsis-icon);
-    }
-
-    :host([variant='mini-compare-chart']) > slot:not([name='icons']) {
-        display: block;
     }
 
     .top-section {
@@ -149,19 +125,6 @@ import{html as o,LitElement as K,nothing as Q}from"/libs/deps/lit-all.min.js";im
         right: initial;
         padding: 8px 11px;
         border-radius: 0 5px 5px 0;
-    }
-
-    .body .catalog-badge {
-        display: flex;
-        height: fit-content;
-        flex-direction: column;
-        width: fit-content;
-        max-width: 140px;
-        border-radius: 5px;
-        position: relative;
-        top: 0;
-        margin-left: var(--consonant-merch-spacing-xxs);
-        box-sizing: border-box;
     }
 
     .detail-bg-container {
@@ -267,60 +230,588 @@ import{html as o,LitElement as K,nothing as Q}from"/libs/deps/lit-all.min.js";im
         margin-top: 2px;
     }
 
-    .twp-badge {
-        padding: 4px 10px 5px 10px;
-    }
-
-    :host([aria-selected]) .twp-badge {
-        margin-inline-end: 2px;
-        padding-inline-end: 9px;
-    }
-
-    :host([variant='twp']) {
-        padding: 4px 10px 5px 10px;
-    }
-
     slot[name='icons'] {
         display: flex;
         gap: 8px;
     }
+`,nt=()=>{let o=[B`
+        /* Tablet */
+        @media screen and ${H(l)} {
+            :host([size='wide']),
+            :host([size='super-wide']) {
+                grid-column: span 3;
+                width: 100%;
+                max-width: var(--consonant-merch-card-tablet-wide-width);
+                margin: 0 auto;
+            }
+        }
 
-    :host([variant='twp']) ::slotted(merch-offer-select) {
-        display: none;
+        /* Laptop */
+        @media screen and ${H(m)} {
+            :host([size='super-wide']) {
+                grid-column: span 3;
+            }
+        `];return o.push(B`
+        /* Large desktop */
+        @media screen and ${H(g)} {
+            :host([size='super-wide']) {
+                grid-column: span 4;
+            }
+        }
+    `),o};import{html as $}from"/libs/deps/lit-all.min.js";var p=class o{static styleMap={};card;constructor(t){if(this.card=t,!o?.styleMap?.[this.card.variant]){o.styleMap[this.card.variant]=!0;let e=document.createElement("style");e.innerHTML=this.getGlobalCSS(),document.head.appendChild(e)}}get badge(){let t;if(!(!this.card.badgeBackgroundColor||!this.card.badgeColor||!this.card.badgeText))return this.evergreen&&(t=`border: 1px solid ${this.card.badgeBackgroundColor}; border-right: none;`),$`
+        <div
+            id="badge"
+            class="${this.card.variant}-badge"
+            style="background-color: ${this.card.badgeBackgroundColor};
+                color: ${this.card.badgeColor};
+                ${t}"
+        >
+            ${this.card.badgeText}
+        </div>
+    `}get cardImage(){return $` <div class="image">
+        <slot name="bg-image"></slot>
+        ${this.badge}
+    </div>`}getGlobalCSS(){return""}get evergreen(){return this.card.classList.contains("intro-pricing")}get promoBottom(){return this.card.classList.contains("promo-bottom")}get secureLabelFooter(){let t=this.card.secureLabel?$`<span class="secure-transaction-label"
+              >${this.card.secureLabel}</span
+          >`:"";return $`<footer>${t}<slot name="footer"></slot></footer>`}get titleSelector(){return'[slot="heading-xs"]'}async adjustTitleWidth(){let t=this.card.getBoundingClientRect().width,e=this.card.badgeElement?.getBoundingClientRect().width||0;t===0||e===0||this.card.style.setProperty("--consonant-merch-card-heading-xs-max-width",`${Math.round(t-e-16)}px`)}postCardUpdateHook(){}connectedCallbackHook(){}disconnectedCallbackHook(){}renderLayout(){}};import{html as U,css as zt}from"/libs/deps/lit-all.min.js";var at=`
+:root {
+  --consonant-merch-card-catalog-width: 276px;
+  --consonant-merch-card-catalog-icon-size: 40px;
+}
+.one-merch-card.catalog,
+.two-merch-cards.catalog,
+.three-merch-cards.catalog,
+.four-merch-cards.catalog {
+    grid-template-columns: var(--consonant-merch-card-catalog-width);
+}
+
+@media screen and ${l} {
+    :root {
+      --consonant-merch-card-catalog-width: 302px;
     }
 
-    :host([variant='twp']) .top-section {
-        flex: 0;
-        display: flex;
-        flex-direction: column;
-        justify-content: flex-start;
-        height: 100%;
-        gap: var(--consonant-merch-spacing-xxs);
-        padding: var(--consonant-merch-spacing-xs)
-            var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-xs)
-            var(--consonant-merch-spacing-xs);
-        align-items: flex-start;
+    .two-merch-cards.catalog,
+    .three-merch-cards.catalog,
+    .four-merch-cards.catalog {
+        grid-template-columns: repeat(2, var(--consonant-merch-card-catalog-width));
+    }
+}
+
+@media screen and ${m} {
+    :root {
+      --consonant-merch-card-catalog-width: 276px;
+    }
+    .three-merch-cards.catalog,
+    .four-merch-cards.catalog {
+        grid-template-columns: repeat(3, var(--consonant-merch-card-catalog-width));
+    }
+}
+
+@media screen and ${g} {
+    .four-merch-cards.catalog {
+        grid-template-columns: repeat(4, var(--consonant-merch-card-catalog-width));
+    }
+}
+
+merch-card[variant="catalog"] [slot="action-menu-content"] {
+  background-color: #000;
+  color: var(--color-white, #fff);
+  font-size: var(--consonant-merch-card-body-xs-font-size);
+  width: fit-content;
+  padding: var(--consonant-merch-spacing-xs);
+  border-radius: var(--consonant-merch-spacing-xxxs);
+  position: absolute;
+  top: 55px;
+  right: 15px;
+  line-height: var(--consonant-merch-card-body-line-height);
+}
+
+merch-card[variant="catalog"] [slot="action-menu-content"] ul {
+  padding-left: 0;
+  padding-bottom: var(--consonant-merch-spacing-xss);
+  margin-top: 0;
+  margin-bottom: 0;
+  list-style-position: inside;
+  list-style-type: '\u2022 ';
+}
+
+merch-card[variant="catalog"] [slot="action-menu-content"] ul li {
+  padding-left: 0;
+  line-height: var(--consonant-merch-card-body-line-height);
+}
+
+merch-card[variant="catalog"] [slot="action-menu-content"] ::marker {
+  margin-right: 0;
+}
+
+merch-card[variant="catalog"] [slot="action-menu-content"] p {
+  color: var(--color-white, #fff);
+}
+
+merch-card[variant="catalog"] [slot="action-menu-content"] a {
+  color: var(--consonant-merch-card-background-color);
+  text-decoration: underline;
+}
+
+merch-card[variant="catalog"] .payment-details {
+  font-size: var(--consonant-merch-card-body-font-size);
+  font-style: italic;
+  font-weight: 400;
+  line-height: var(--consonant-merch-card-body-line-height);
+}`;var b=class extends p{constructor(t){super(t)}renderLayout(){return U` <div class="body">
+        <div class="top-section">
+            <slot name="icons"></slot> ${this.badge}
+            <div
+                class="action-menu
+                ${X()&&this.card.actionMenu?"always-visible":""}
+                ${this.card.actionMenu?"invisible":"hidden"}"
+                @click="${this.card.toggleActionMenu}"
+            ></div>
+        </div>
+        <slot
+            name="action-menu-content"
+            class="action-menu-content
+            ${this.card.actionMenuContent?"":"hidden"}"
+            >${this.card.actionMenuContent}</slot
+        >
+        <slot name="heading-xs"></slot>
+        <slot name="heading-m"></slot>
+        <slot name="body-xxs"></slot>
+        ${this.promoBottom?"":U`<slot name="promo-text"></slot
+                  ><slot name="callout-content"></slot>`}
+        <slot name="body-xs"></slot>
+        ${this.promoBottom?U`<slot name="promo-text"></slot
+                  ><slot name="callout-content"></slot>`:""}
+    </div>
+    ${this.secureLabelFooter}`}toggleActionMenu(t){let e=t?.type==="mouseleave"?!0:void 0,r=this.shadowRoot.querySelector('slot[name="action-menu-content"]');r&&(e||this.card.dispatchEvent(new CustomEvent(Q,{bubbles:!0,composed:!0,detail:{card:this.name,type:"action-menu"}})),r.classList.toggle("hidden",e))}getGlobalCSS(){return at}connectedCallbackHook(){this.card.addEventListener("mouseleave",this.toggleActionMenu)}disconnectedCallbackHook(){this.card.removeEventListener("mouseleave",this.toggleActionMenu)}static variantStyle=zt`
+    :host([variant='catalog']) {
+      min-height: 330px;
     }
 
-    :host([variant='twp']) .body {
-        padding: 0 var(--consonant-merch-spacing-xs);
+    .body .catalog-badge {
+      display: flex;
+      height: fit-content;
+      flex-direction: column;
+      width: fit-content;
+      max-width: 140px;
+      border-radius: 5px;
+      position: relative;
+      top: 0;
+      margin-left: var(--consonant-merch-spacing-xxs);
+      box-sizing: border-box;
     }
+  `};import{html as V,css as Tt}from"/libs/deps/lit-all.min.js";var ct=`
+:root {
+  --consonant-merch-card-ccd-action-width: 276px;
+  --consonant-merch-card-ccd-action-min-height: 320px;
+}
 
-    :host([variant='twp']) footer {
-        gap: var(--consonant-merch-spacing-xxs);
-        flex-direction: column;
-        align-self: flex-start;
+.one-merch-card.ccd-action,
+.two-merch-cards.ccd-action,
+.three-merch-cards.ccd-action,
+.four-merch-cards.ccd-action {
+    grid-template-columns: var(--consonant-merch-card-ccd-action-width);
+}
+
+merch-card[variant="ccd-action"] .price-strikethrough {
+    font-size: 18px;
+}
+
+@media screen and ${l} {
+  .two-merch-cards.ccd-action,
+  .three-merch-cards.ccd-action,
+  .four-merch-cards.ccd-action {
+      grid-template-columns: repeat(2, var(--consonant-merch-card-ccd-action-width));
+  }
+}
+
+@media screen and ${m} {
+  .three-merch-cards.ccd-action,
+  .four-merch-cards.ccd-action {
+      grid-template-columns: repeat(3, var(--consonant-merch-card-ccd-action-width));
+  }
+}
+
+@media screen and ${g} {
+  .four-merch-cards.ccd-action {
+      grid-template-columns: repeat(4, var(--consonant-merch-card-ccd-action-width));
+  }
+}
+`;var y=class extends p{constructor(t){super(t)}getGlobalCSS(){return ct}renderLayout(){return V` <div class="body">
+        <slot name="icons"></slot> ${this.badge}
+        <slot name="heading-xs"></slot>
+        <slot name="heading-m"></slot>
+        ${this.promoBottom?V`<slot name="body-xs"></slot><slot name="promo-text"></slot>`:V`<slot name="promo-text"></slot><slot name="body-xs"></slot>`}
+        <footer><slot name="footer"></slot></footer>
+        ${this.defaultSlot}
+    </div>`}static variantStyle=Tt`
+    :host([variant='ccd-action']:not([size])) {
+      width: var(--consonant-merch-card-ccd-action-width);
     }
+  `};import{html as w}from"/libs/deps/lit-all.min.js";var it=`
+:root {
+  --consonant-merch-card-image-width: 300px;
+  --consonant-merch-card-image-height: 180px;
+}
 
-    :host([variant='special-offers'].center) {
-        text-align: center;
+.one-merch-card.image,
+.two-merch-cards.image,
+.three-merch-cards.image,
+.four-merch-cards.image {
+  grid-template-columns: var(--consonant-merch-card-image-width);
+}
+
+@media screen and ${l} {
+  .two-merch-cards.image,
+  .three-merch-cards.image,
+  .four-merch-cards.image {
+      grid-template-columns: repeat(2, var(--consonant-merch-card-image-width));
+  }
+}
+
+@media screen and ${m} {
+  :root {
+    --consonant-merch-card-image-width: 378px;
+  }
+  .three-merch-cards.image,
+  .four-merch-cards.image {
+      grid-template-columns: repeat(3, var(--consonant-merch-card-image-width));
+  }
+}
+
+@media screen and ${g} {
+  .four-merch-cards.image {
+      grid-template-columns: repeat(4, var(--consonant-merch-card-image-width));
+  }
+}
+`;var A=class extends p{constructor(t){super(t)}getGlobalCSS(){return it}renderLayout(){return w`${this.cardImage}
+    <div class="body">
+        <slot name="icons"></slot>
+        <slot name="heading-xs"></slot>
+        <slot name="body-xxs"></slot>
+        ${this.promoBottom?w`<slot name="body-xs"></slot><slot name="promo-text"></slot>`:w`<slot name="promo-text"></slot><slot name="body-xs"></slot>`}
+    </div>
+    ${this.evergreen?w`
+              <div
+                  class="detail-bg-container"
+                  style="background: ${this.card.detailBg}"
+              >
+                  <slot name="detail-bg"></slot>
+              </div>
+          `:w`
+              <hr />
+              ${this.secureLabelFooter}
+          `}`}};import{html as ht}from"/libs/deps/lit-all.min.js";var st=`
+--consonant-merch-card-inline-heading-width: 300px;
+
+.one-merch-card.inline-heading,
+.two-merch-cards.inline-heading,
+.three-merch-cards.inline-heading,
+.four-merch-cards.inline-heading {
+    grid-template-columns: var(--consonant-merch-card-inline-heading-width);
+}
+
+@media screen and ${l} {
+  .two-merch-cards.inline-heading,
+  .three-merch-cards.inline-heading,
+  .four-merch-cards.inline-heading {
+      grid-template-columns: repeat(2, var(--consonant-merch-card-inline-heading-width));
+  }
+}
+
+@media screen and ${m} {
+  --consonant-merch-card-inline-heading-width: 378px;
+
+  .three-merch-cards.inline-heading,
+  .four-merch-cards.inline-heading {
+      grid-template-columns: repeat(3, var(--consonant-merch-card-inline-heading-width));
+  }
+}
+
+@media screen and ${g} {
+  .four-merch-cards.inline-heading {
+      grid-template-columns: repeat(4, var(--consonant-merch-card-inline-heading-width));
+  }
+}
+`;var M=class extends p{constructor(t){super(t)}getGlobalCSS(){return st}renderLayout(){return ht` ${this.badge}
+    <div class="body">
+        <div class="top-section">
+            <slot name="icons"></slot>
+            <slot name="heading-xs"></slot>
+        </div>
+        <slot name="body-xs"></slot>
+    </div>
+    ${this.card.customHr?"":ht`<hr />`} ${this.secureLabelFooter}`}};import{html as O,css as Lt,unsafeCSS as mt}from"/libs/deps/lit-all.min.js";var dt=`
+  :root {
+    --consonant-merch-card-mini-compare-chart-icon-size: 32px;
+    --consonant-merch-card-mini-compare-mobile-cta-font-size: 15px;
+    --consonant-merch-card-mini-compare-mobile-cta-width: 75px;
+    --consonant-merch-card-mini-compare-badge-mobile-max-width: 50px;
+  }
+  
+  merch-card[variant="mini-compare-chart"] [slot="heading-m"] {
+    padding: 0 var(--consonant-merch-spacing-s) 0;
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot="body-m"] {
+    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s);
+  }
+
+  merch-card[variant="mini-compare-chart"] [is="inline-price"] {
+    display: inline-block;
+    min-height: 30px;
+    min-width: 1px;
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot='callout-content'] {
+    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s) 0px;
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot='callout-content'] [is="inline-price"] {
+    min-height: unset;
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot="price-commitment"] {
+    font-size: var(--consonant-merch-card-body-xs-font-size);
+    padding: 0 var(--consonant-merch-spacing-s);
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot="price-commitment"] a {
+    display: inline-block;
+    height: 27px;
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot="offers"] {
+    font-size: var(--consonant-merch-card-body-xs-font-size);
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot="body-xxs"] {
+    font-size: var(--consonant-merch-card-body-xs-font-size);
+    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s) 0;    
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot="promo-text"] {
+    font-size: var(--consonant-merch-card-body-m-font-size);
+    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s) 0;
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot="promo-text"] a {
+    text-decoration: underline;
+  }
+
+  merch-card[variant="mini-compare-chart"] .footer-row-icon {
+    display: flex;
+    place-items: center;
+  }
+
+  merch-card[variant="mini-compare-chart"] .footer-row-icon img {
+    max-width: initial;
+    width: var(--consonant-merch-card-mini-compare-chart-icon-size);
+    height: var(--consonant-merch-card-mini-compare-chart-icon-size);
+  }
+
+  merch-card[variant="mini-compare-chart"] .footer-row-cell {
+    border-top: 1px solid var(--consonant-merch-card-border-color);
+    display: flex;
+    gap: var(--consonant-merch-spacing-xs);
+    justify-content: start;
+    place-items: center;
+    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s);
+  }
+
+  merch-card[variant="mini-compare-chart"] .footer-row-cell-description {
+    font-size: var(--consonant-merch-card-body-s-font-size);
+    line-height: var(--consonant-merch-card-body-s-line-height);
+  }
+
+  merch-card[variant="mini-compare-chart"] .footer-row-cell-description p {
+    color: var(--merch-color-grey-80);
+    vertical-align: bottom;
+  }
+
+  merch-card[variant="mini-compare-chart"] .footer-row-cell-description a {
+    color: var(--color-accent);
+    text-decoration: solid;
+  }
+  
+.one-merch-card.mini-compare-chart {
+  grid-template-columns: var(--consonant-merch-card-mini-compare-chart-wide-width);
+  gap: var(--consonant-merch-spacing-xs);
+}
+
+.two-merch-cards.mini-compare-chart,
+.three-merch-cards.mini-compare-chart,
+.four-merch-cards.mini-compare-chart {
+  grid-template-columns: repeat(2, var(--consonant-merch-card-mini-compare-chart-width));
+  gap: var(--consonant-merch-spacing-xs);
+}
+
+/* mini compare mobile */ 
+@media screen and ${f} {
+  :root {
+    --consonant-merch-card-mini-compare-chart-width: 302px;
+    --consonant-merch-card-mini-compare-chart-wide-width: 302px;
+  }
+
+  .two-merch-cards.mini-compare-chart,
+  .three-merch-cards.mini-compare-chart,
+  .four-merch-cards.mini-compare-chart {
+    grid-template-columns: var(--consonant-merch-card-mini-compare-chart-width);
+    gap: var(--consonant-merch-spacing-xs);
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot='heading-m'] {
+    font-size: var(--consonant-merch-card-body-s-font-size);
+    line-height: var(--consonant-merch-card-body-s-line-height);
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot='heading-m-price'] {
+    font-size: var(--consonant-merch-card-body-s-font-size);
+    line-height: var(--consonant-merch-card-body-s-line-height);
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot='body-m'] {
+    font-size: var(--consonant-merch-card-body-xs-font-size);
+    line-height: var(--consonant-merch-card-body-xs-line-height);
+  }
+  
+  merch-card[variant="mini-compare-chart"] [slot="promo-text"] {
+    font-size: var(--consonant-merch-card-body-xs-font-size);
+    line-height: var(--consonant-merch-card-body-xs-line-height);
+  }
+  merch-card[variant="mini-compare-chart"] .footer-row-cell-description {
+    font-size: var(--consonant-merch-card-body-xs-font-size);
+    line-height: var(--consonant-merch-card-body-xs-line-height);
+  }
+}
+
+@media screen and ${_} {
+  .three-merch-cards.mini-compare-chart merch-card [slot="footer"] a,
+  .four-merch-cards.mini-compare-chart merch-card [slot="footer"] a {
+    flex: 1;
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot='heading-m'] {
+    font-size: var(--consonant-merch-card-body-s-font-size);
+    line-height: var(--consonant-merch-card-body-s-line-height);
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot='heading-m-price'] {
+    font-size: var(--consonant-merch-card-body-s-font-size);
+    line-height: var(--consonant-merch-card-body-s-line-height);
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot='body-m'] {
+    font-size: var(--consonant-merch-card-body-xs-font-size);
+    line-height: var(--consonant-merch-card-body-xs-line-height);
+  }
+
+  merch-card[variant="mini-compare-chart"] [slot="promo-text"] {
+    font-size: var(--consonant-merch-card-body-xs-font-size);
+    line-height: var(--consonant-merch-card-body-xs-line-height);
+  }
+  
+  merch-card[variant="mini-compare-chart"] .footer-row-cell-description {
+    font-size: var(--consonant-merch-card-body-xs-font-size);
+    line-height: var(--consonant-merch-card-body-xs-line-height);
+  }
+}
+@media screen and ${l} {
+  :root {
+    --consonant-merch-card-mini-compare-chart-width: 302px;
+    --consonant-merch-card-mini-compare-chart-wide-width: 302px;
+  }
+
+  .two-merch-cards.mini-compare-chart {
+    grid-template-columns: repeat(2, minmax(var(--consonant-merch-card-mini-compare-chart-width), var(--consonant-merch-card-mini-compare-chart-wide-width)));
+    gap: var(--consonant-merch-spacing-m);
+  }
+
+  .three-merch-cards.mini-compare-chart,
+  .four-merch-cards.mini-compare-chart {
+      grid-template-columns: repeat(2, minmax(var(--consonant-merch-card-mini-compare-chart-width), var(--consonant-merch-card-mini-compare-chart-wide-width)));
+  }
+}
+
+/* desktop */
+@media screen and ${m} {
+  :root {
+    --consonant-merch-card-mini-compare-chart-width: 378px;
+    --consonant-merch-card-mini-compare-chart-wide-width: 484px;  
+  }
+  .one-merch-card.mini-compare-chart {
+    grid-template-columns: var(--consonant-merch-card-mini-compare-chart-wide-width);
+  }
+
+  .two-merch-cards.mini-compare-chart {
+    grid-template-columns: repeat(2, var(--consonant-merch-card-mini-compare-chart-wide-width));
+    gap: var(--consonant-merch-spacing-m);
+  }
+
+  .three-merch-cards.mini-compare-chart,
+  .four-merch-cards.mini-compare-chart {
+    grid-template-columns: repeat(3, var(--consonant-merch-card-mini-compare-chart-width));
+    gap: var(--consonant-merch-spacing-m);
+  }
+}
+
+@media screen and ${g} {
+  .four-merch-cards.mini-compare-chart {
+      grid-template-columns: repeat(4, var(--consonant-merch-card-mini-compare-chart-width));
+  }
+}
+
+merch-card .footer-row-cell:nth-child(1) {
+  min-height: var(--consonant-merch-card-footer-row-1-min-height);
+}
+
+merch-card .footer-row-cell:nth-child(2) {
+  min-height: var(--consonant-merch-card-footer-row-2-min-height);
+}
+
+merch-card .footer-row-cell:nth-child(3) {
+  min-height: var(--consonant-merch-card-footer-row-3-min-height);
+}
+
+merch-card .footer-row-cell:nth-child(4) {
+  min-height: var(--consonant-merch-card-footer-row-4-min-height);
+}
+
+merch-card .footer-row-cell:nth-child(5) {
+  min-height: var(--consonant-merch-card-footer-row-5-min-height);
+}
+
+merch-card .footer-row-cell:nth-child(6) {
+  min-height: var(--consonant-merch-card-footer-row-6-min-height);
+}
+
+merch-card .footer-row-cell:nth-child(7) {
+  min-height: var(--consonant-merch-card-footer-row-7-min-height);
+}
+
+merch-card .footer-row-cell:nth-child(8) {
+  min-height: var(--consonant-merch-card-footer-row-8-min-height);
+}
+`;var _t=32,E=class extends p{constructor(t){super(t)}#t;getRowMinHeightPropertyName=t=>`--consonant-merch-card-footer-row-${t}-min-height`;getContainer(){return this.#t=this.#t??this.card.closest('[class*="-merch-cards"]')??this.card.parentElement,this.#t}getGlobalCSS(){return dt}getMiniCompareFooter=()=>{let t=this.card.secureLabel?O`<slot name="secure-transaction-label">
+              <span class="secure-transaction-label"
+                  >${this.card.secureLabel}</span
+              ></slot
+          >`:O`<slot name="secure-transaction-label"></slot>`;return O`<footer>${t}<slot name="footer"></slot></footer>`};updateMiniCompareElementMinHeight(t,e){let r=`--consonant-merch-card-mini-compare-${e}-height`,n=Math.max(0,parseInt(window.getComputedStyle(t).height)||0),d=parseInt(this.getContainer().style.getPropertyValue(r))||0;n>d&&this.getContainer().style.setProperty(r,`${n}px`)}adjustMiniCompareBodySlots(){if(this.card.getBoundingClientRect().width===0)return;this.updateMiniCompareElementMinHeight(this.card.shadowRoot.querySelector(".top-section"),"top-section"),["heading-m","body-m","heading-m-price","price-commitment","offers","promo-text","callout-content","secure-transaction-label"].forEach(r=>this.updateMiniCompareElementMinHeight(this.card.shadowRoot.querySelector(`slot[name="${r}"]`),r)),this.updateMiniCompareElementMinHeight(this.card.shadowRoot.querySelector("footer"),"footer");let e=this.card.shadowRoot.querySelector(".mini-compare-chart-badge");e&&e.textContent!==""&&this.getContainer().style.setProperty("--consonant-merch-card-mini-compare-top-section-mobile-height","32px")}adjustMiniCompareFooterRows(){if(this.card.getBoundingClientRect().width===0)return;[...this.card.querySelector('[slot="footer-rows"]').children].forEach((e,r)=>{let n=Math.max(_t,parseInt(window.getComputedStyle(e).height)||0),d=parseInt(this.getContainer().style.getPropertyValue(this.getRowMinHeightPropertyName(r+1)))||0;n>d&&this.getContainer().style.setProperty(this.getRowMinHeightPropertyName(r+1),`${n}px`)})}removeEmptyRows(){this.card.querySelectorAll(".footer-row-cell").forEach(e=>{let r=e.querySelector(".footer-row-cell-description");r&&!r.textContent.trim()&&e.remove()})}renderLayout(){return O` <div class="top-section${this.badge?" badge":""}">
+            <slot name="icons"></slot> ${this.badge}
+        </div>
+        <slot name="heading-m"></slot>
+        <slot name="body-m"></slot>
+        <slot name="heading-m-price"></slot>
+        <slot name="body-xxs"></slot>
+        <slot name="price-commitment"></slot>
+        <slot name="offers"></slot>
+        <slot name="promo-text"></slot>
+        <slot name="callout-content"></slot>
+        ${this.getMiniCompareFooter()}
+        <slot name="footer-rows"><slot name="body-s"></slot></slot>`}postCardUpdateHook(){J()?this.removeEmptyRows():(this.adjustMiniCompareBodySlots(),this.adjustMiniCompareFooterRows())}static variantStyle=Lt`
+    :host([variant='mini-compare-chart']) > slot:not([name='icons']) {
+        display: block;
     }
-
-    /* plans */
-    :host([variant='plans']) {
-        min-height: 348px;
-    }
-
     :host([variant='mini-compare-chart']) footer {
         min-height: var(--consonant-merch-card-mini-compare-footer-height);
         padding: var(--consonant-merch-spacing-xs);
@@ -333,7 +824,7 @@ import{html as o,LitElement as K,nothing as Q}from"/libs/deps/lit-all.min.js";im
         height: var(--consonant-merch-card-mini-compare-top-section-height);
     }
 
-    @media screen and ${y(b)} {
+    @media screen and ${mt(_)} {
         [class*'-merch-cards'] :host([variant='mini-compare-chart']) footer {
             flex-direction: column;
             align-items: stretch;
@@ -341,7 +832,7 @@ import{html as o,LitElement as K,nothing as Q}from"/libs/deps/lit-all.min.js";im
         }
     }
 
-    @media screen and ${y(s)} {
+    @media screen and ${mt(m)} {
         :host([variant='mini-compare-chart']) footer {
             padding: var(--consonant-merch-spacing-xs)
                 var(--consonant-merch-spacing-s)
@@ -384,36 +875,407 @@ import{html as o,LitElement as K,nothing as Q}from"/libs/deps/lit-all.min.js";im
             --consonant-merch-card-mini-compare-callout-content-height
         );
     }
+  `};import{html as R,css as $t}from"/libs/deps/lit-all.min.js";var lt=`
+--consonant-merch-card-plans-width: 300px;
+--consonant-merch-card-plans-icon-size: 40px;
+merch-card[variant="plans"] [slot="description"] {
+  min-height: 84px;
+}
 
-    :host([variant='plans']) ::slotted([slot='heading-xs']),
-    :host([variant='segment']) ::slotted([slot='heading-xs']) {
-        max-width: var(--consonant-merch-card-heading-xs-max-width, 100%);
+merch-card[variant="plans"] [slot="quantity-select"] {
+  display: flex;
+  justify-content: flex-start;
+  box-sizing: border-box;
+  width: 100%;
+  padding: var(--consonant-merch-spacing-xs);
+}
+
+.one-merch-card.plans,
+.two-merch-cards.plans,
+.three-merch-cards.plans,
+.four-merch-cards.plans {
+    grid-template-columns: var(--consonant-merch-card-plans-width);
+}
+
+/* Tablet */
+@media screen and ${l} {
+  :root {
+    --consonant-merch-card-plans-width: 302px;
+  }
+  .two-merch-cards.plans,
+  .three-merch-cards.plans,
+  .four-merch-cards.plans {
+      grid-template-columns: repeat(2, var(--consonant-merch-card-plans-width));
+  }
+}
+
+/* desktop */
+@media screen and ${m} {
+  :root {
+    --consonant-merch-card-plans-width: 276px;
+  }
+  .three-merch-cards.plans,
+  .four-merch-cards.plans {
+      grid-template-columns: repeat(3, var(--consonant-merch-card-plans-width));
+  }
+}
+
+/* Large desktop */
+    @media screen and ${g} {
+    .four-merch-cards.plans {
+        grid-template-columns: repeat(4, var(--consonant-merch-card-plans-width));
     }
-`,I=()=>{let c=[S`
-        /* Tablet */
-        @media screen and ${y(h)} {
-            :host([size='wide']),
-            :host([size='super-wide']) {
-                grid-column: span 3;
-                width: 100%;
-                max-width: var(--consonant-merch-card-tablet-wide-width);
-                margin: 0 auto;
-            }
-        }
+}
+`;var S=class extends p{constructor(t){super(t)}getGlobalCSS(){return lt}postCardUpdateHook(){this.adjustTitleWidth()}get stockCheckbox(){return this.checkboxLabel?R`<label id="stock-checkbox">
+                <input type="checkbox" @change=${this.card.toggleStockOffer}></input>
+                <span></span>
+                ${this.card.checkboxLabel}
+            </label>`:""}renderLayout(){return R` ${this.badge}
+        <div class="body">
+            <slot name="icons"></slot>
+            <slot name="heading-xs"></slot>
+            <slot name="heading-m"></slot>
+            <slot name="body-xxs"></slot>
+            ${this.promoBottom?"":R`<slot name="promo-text"></slot><slot name="callout-content"></slot> `}
+            <slot name="body-xs"></slot>
+            ${this.promoBottom?R`<slot name="promo-text"></slot><slot name="callout-content"></slot> `:""}  
+            ${this.stockCheckbox}
+        </div>
+        <slot name="quantity-select"></slot>
+        ${this.secureLabelFooter}`}static variantStyle=$t`
+    :host([variant='plans']) {
+      min-height: 348px;
+    }
+    :host([variant='plans']) ::slotted([slot='heading-xs']),
+      max-width: var(--consonant-merch-card-heading-xs-max-width, 100%);
+    }
+  `};import{html as G}from"/libs/deps/lit-all.min.js";var pt=`
+:root {
+  --consonant-merch-card-product-width: 300px;
+}
 
-        /* Laptop */
-        @media screen and ${y(s)} {
-            :host([size='super-wide']) {
-                grid-column: span 3;
-            }
-        `];return c.push(S`
-        /* Large desktop */
-        @media screen and ${y(d)} {
-            :host([size='super-wide']) {
-                grid-column: span 4;
-            }
-        }
-    `),c};var q=document.createElement("style");q.innerHTML=`
+/* grid style for product */
+.one-merch-card.product,
+.two-merch-cards.product,
+.three-merch-cards.product,
+.four-merch-cards.product {
+    grid-template-columns: var(--consonant-merch-card-product-width);
+}
+
+/* Tablet */
+@media screen and ${l} {
+    .two-merch-cards.product,
+    .three-merch-cards.product,
+    .four-merch-cards.product {
+        grid-template-columns: repeat(2, var(--consonant-merch-card-product-width));
+    }
+}
+
+/* desktop */
+@media screen and ${m} {
+  :root {
+    --consonant-merch-card-product-width: 378px;
+  }
+  .three-merch-cards.product,
+  .four-merch-cards.product {
+      grid-template-columns: repeat(3, var(--consonant-merch-card-product-width));
+  }
+}
+
+/* Large desktop */
+@media screen and ${g} {
+  .four-merch-cards.product {
+      grid-template-columns: repeat(4, var(--consonant-merch-card-product-width));
+  }
+}
+`;var C=class extends p{constructor(t){super(t)}getGlobalCSS(){return pt}renderLayout(){return G` ${this.badge}
+      <div class="body">
+          <slot name="icons"></slot>
+          <slot name="heading-xs"></slot>
+          <slot name="body-xxs"></slot>
+          ${this.promoBottom?"":G`<slot name="promo-text"></slot><slot name="callout-content"></slot>`}
+          <slot name="body-xs"></slot>
+          ${this.promoBottom?G`<slot name="promo-text"></slot><slot name="callout-content"></slot>`:""}
+      </div>
+      ${this.secureLabelFooter}`}};import{html as q,css as At}from"/libs/deps/lit-all.min.js";var gt=`
+:root {
+  --consonant-merch-card-segment-width: 378px;
+}
+
+/* grid style for segment */
+.one-merch-card.segment,
+.two-merch-cards.segment,
+.three-merch-cards.segment,
+.four-merch-cards.segment {
+  grid-template-columns: minmax(276px, var(--consonant-merch-card-segment-width));
+}
+
+/* Mobile */
+@media screen and ${f} {
+  :root {
+    --consonant-merch-card-segment-width: 276px;
+  }
+}
+
+@media screen and ${l} {
+  :root {
+    --consonant-merch-card-segment-width: 276px;
+  }
+  .two-merch-cards.segment,
+  .three-merch-cards.segment,
+  .four-merch-cards.segment {
+      grid-template-columns: repeat(2, minmax(276px, var(--consonant-merch-card-segment-width)));
+  }
+}
+
+/* desktop */
+@media screen and ${m} {
+  :root {
+    --consonant-merch-card-segment-width: 302px;
+  }
+  .three-merch-cards.segment {
+      grid-template-columns: repeat(3, minmax(276px, var(--consonant-merch-card-segment-width)));
+  }
+
+  .four-merch-cards.segment {
+      grid-template-columns: repeat(4, minmax(276px, var(--consonant-merch-card-segment-width)));
+  }
+}
+`;var k=class extends p{constructor(t){super(t)}getGlobalCSS(){return gt}postCardUpdateHook(){this.adjustTitleWidth()}renderLayout(){return q` ${this.badge}
+    <div class="body">
+        <slot name="heading-xs"></slot>
+        <slot name="body-xxs"></slot>
+        ${this.promoBottom?"":q`<slot name="promo-text"></slot><slot name="callout-content"></slot>`}
+        <slot name="body-xs"></slot>
+        ${this.promoBottom?q`<slot name="promo-text"></slot><slot name="callout-content"></slot>`:""}
+    </div>
+    <hr />
+    ${this.secureLabelFooter}`}static variantStyle=At`
+    :host([variant='segment']) {
+      min-height: 214px;
+    }
+    :host([variant='segment']) ::slotted([slot='heading-xs']) {
+      max-width: var(--consonant-merch-card-heading-xs-max-width, 100%);
+    }
+  `};import{html as j,css as Mt}from"/libs/deps/lit-all.min.js";var ut=`
+:root {
+  --consonant-merch-card-special-offers-width: 378px;
+}
+
+merch-card[variant="special-offers"] span[is="inline-price"][data-template="strikethrough"] {
+  font-size: var(--consonant-merch-card-body-xs-font-size);
+}
+
+/* grid style for special-offers */
+.one-merch-card.special-offers,
+.two-merch-cards.special-offers,
+.three-merch-cards.special-offers,
+.four-merch-cards.special-offers {
+  grid-template-columns: minmax(300px, var(--consonant-merch-card-special-offers-width));
+}
+
+@media screen and ${f} {
+  :root\xA0{
+    --consonant-merch-card-special-offers-width: 302px;
+  }
+} 
+  
+@media screen and ${l} {
+  :root {
+    --consonant-merch-card-special-offers-width: 302px;
+  }
+  .two-merch-cards.special-offers,
+  .three-merch-cards.special-offers,
+  .four-merch-cards.special-offers {
+      grid-template-columns: repeat(2, minmax(300px, var(--consonant-merch-card-special-offers-width)));
+  }
+}
+
+/* desktop */
+@media screen and ${m} {
+  .three-merch-cards.special-offers,
+  .four-merch-cards.special-offers {
+    grid-template-columns: repeat(3, minmax(300px, var(--consonant-merch-card-special-offers-width)));
+  }
+}
+
+@media screen and ${g} {
+  .four-merch-cards.special-offers {
+    grid-template-columns: repeat(4, minmax(300px, var(--consonant-merch-card-special-offers-width)));
+  }
+}
+`;var z=class extends p{constructor(t){super(t)}getGlobalCSS(){return ut}renderLayout(){return j`${this.cardImage}
+          <div class="body">
+              <slot name="detail-m"></slot>
+              <slot name="heading-xs"></slot>
+              <slot name="body-xs"></slot>
+          </div>
+          ${this.evergreen?j`
+                    <div
+                        class="detail-bg-container"
+                        style="background: ${this.card.detailBg}"
+                    >
+                        <slot name="detail-bg"></slot>
+                    </div>
+                `:j`
+                    <hr />
+                    ${this.secureLabelFooter}
+                `}`}get titleSelector(){return'[slot="detail-m"]'}static variantStyle=Mt`
+    :host([variant='special-offers']) {
+      min-height: 439px;
+    }
+    :host([variant='special-offers'].center) {
+      text-align: center;
+    }  
+  `};import{html as Ot,css as Rt}from"/libs/deps/lit-all.min.js";var ft=`
+:root {
+  --consonant-merch-card-twp-width: 268px;
+  --consonant-merch-card-twp-mobile-width: 300px;
+  --consonant-merch-card-twp-mobile-height: 358px;
+}
+merch-card[variant="twp"] div[class$='twp-badge'] {
+  padding: 4px 10px 5px 10px;
+}
+
+merch-card[variant="twp"] [slot="body-xs-top"] {
+  font-size: var(--consonant-merch-card-body-xs-font-size);
+  line-height: var(--consonant-merch-card-body-xs-line-height);
+  color: var(--merch-color-grey-80);
+}
+
+merch-card[variant="twp"] [slot="body-xs"] ul {
+  padding: 0;
+  margin: 0;
+}
+
+merch-card[variant="twp"] [slot="body-xs"] ul li {
+  list-style-type: none;
+  padding-left: 0;
+}
+
+merch-card[variant="twp"] [slot="body-xs"] ul li::before {
+  content: '\xB7';
+  font-size: 20px;
+  padding-right: 5px;
+  font-weight: bold;
+}
+
+merch-card[variant="twp"] [slot="footer"] {
+  font-size: var(--consonant-merch-card-body-xs-font-size);
+  line-height: var(--consonant-merch-card-body-xs-line-height);
+  padding: var(--consonant-merch-spacing-s)
+  var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-xs);
+  color: var(--merch-color-grey-80);
+  display: flex;
+  flex-flow: wrap;
+}
+
+merch-card[variant='twp'] merch-quantity-select,
+merch-card[variant='twp'] merch-offer-select {
+  display: none;
+}
+
+.one-merch-card.twp,
+.two-merch-cards.twp,
+.three-merch-cards.twp {
+  grid-template-columns: var(--consonant-merch-card-image-width);
+}
+
+@media screen and ${m} {
+  :root {
+    --consonant-merch-card-twp-width: 268px;
+  }
+  .one-merch-card.twp
+  .two-merch-cards.twp {
+      grid-template-columns: repeat(2, var(--consonant-merch-card-twp-width));
+  }
+  .three-merch-cards.twp {
+      grid-template-columns: repeat(3, var(--consonant-merch-card-twp-width));
+  }
+}
+
+@media screen and ${g} {
+    .one-merch-card.twp
+    .two-merch-cards.twp {
+        grid-template-columns: repeat(2, var(--consonant-merch-card-twp-width));
+    }
+    .three-merch-cards.twp {
+    grid-template-columns: repeat(3, var(--consonant-merch-card-twp-width));
+  }
+}
+
+@media screen and ${f} {
+  :root {
+    --consonant-merch-card-twp-width: 300px;
+  }
+  .one-merch-card.twp,
+  .two-merch-cards.twp,
+  .three-merch-cards.twp {
+      grid-template-columns: repeat(1, var(--consonant-merch-card-twp-mobile-width));
+  }
+}
+
+@media screen and ${l} {
+  :root {
+    --consonant-merch-card-twp-width: 268px;
+  }
+  .one-merch-card.twp,
+  .two-merch-cards.twp {
+      grid-template-columns: repeat(2, var(--consonant-merch-card-twp-width));
+  }
+  .three-merch-cards.twp {
+      grid-template-columns: repeat(3, var(--consonant-merch-card-twp-width));
+  }
+}
+`;var T=class extends p{constructor(t){super(t)}getGlobalCSS(){return ft}renderLayout(){return Ot`${this.badge}
+      <div class="top-section">
+          <slot name="icons"></slot>
+          <slot name="heading-xs"></slot>
+          <slot name="body-xs-top"></slot>
+      </div>
+      <div class="body">
+          <slot name="body-xs"></slot>
+      </div>
+      <footer><slot name="footer"></slot></footer>`}static variantStyle=Rt`
+    :host([variant='twp']) {
+      padding: 4px 10px 5px 10px;
+    }
+    .twp-badge {
+      padding: 4px 10px 5px 10px;
+    }
+
+    :host([variant='twp']) ::slotted(merch-offer-select) {
+      display: none;
+    }
+
+    :host([variant='twp']) .top-section {
+      flex: 0;
+      display: flex;
+      flex-direction: column;
+      justify-content: flex-start;
+      height: 100%;
+      gap: var(--consonant-merch-spacing-xxs);
+      padding: var(--consonant-merch-spacing-xs)
+          var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-xs)
+          var(--consonant-merch-spacing-xs);
+      align-items: flex-start;
+    }
+
+    :host([variant='twp']) .body {
+      padding: 0 var(--consonant-merch-spacing-xs);
+    }
+    
+    :host([aria-selected]) .twp-badge {
+        margin-inline-end: 2px;
+        padding-inline-end: 9px;
+    }
+
+    :host([variant='twp']) footer {
+      gap: var(--consonant-merch-spacing-xxs);
+      flex-direction: column;
+      align-self: flex-start;
+    }
+  `};var xt=o=>{switch(o.variant){case"catalog":return new b(o);case"ccd-action":return new y(o);case"image":return new A(o);case"inline-heading":return new M(o);case"mini-compare-chart":return new E(o);case"plans":return new S(o);case"product":return new C(o);case"segment":return new k(o);case"special-offers":return new z(o);case"twp":return new T(o);default:return new C(o)}},vt=()=>{let o=[];return o.push(b.variantStyle),o.push(y.variantStyle),o.push(E.variantStyle),o.push(S.variantStyle),o.push(k.variantStyle),o.push(z.variantStyle),o.push(T.variantStyle),o};var bt=document.createElement("style");bt.innerHTML=`
 :root {
     --consonant-merch-card-detail-font-size: 12px;
     --consonant-merch-card-detail-font-weight: 500;
@@ -476,7 +1338,6 @@ import{html as o,LitElement as K,nothing as Q}from"/libs/deps/lit-all.min.js";im
 
 
     --consonant-merch-card-heading-padding: 0;
-    --consonant-merch-card-image-height: 180px;
 
     /* colors */
     --consonant-merch-card-border-color: #eaeaea;
@@ -488,45 +1349,6 @@ import{html as o,LitElement as K,nothing as Q}from"/libs/deps/lit-all.min.js";im
     /* merch card generic */
     --consonant-merch-card-max-width: 300px;
     --transition: cmax-height 0.3s linear, opacity 0.3s linear;
-
-    /* special offers */
-    --consonant-merch-card-special-offers-width: 378px;
-
-    /* image */
-    --consonant-merch-card-image-width: 300px;
-
-    /* segment */
-    --consonant-merch-card-segment-width: 378px;
-
-    /* inline-heading */
-    --consonant-merch-card-inline-heading-width: 300px;
-
-    /* product */
-    --consonant-merch-card-product-width: 300px;
-
-    /* plans */
-    --consonant-merch-card-plans-width: 300px;
-    --consonant-merch-card-plans-icon-size: 40px;
-
-    /* catalog */
-    --consonant-merch-card-catalog-width: 276px;
-    --consonant-merch-card-catalog-icon-size: 40px;
-
-    /* twp */
-    --consonant-merch-card-twp-width: 268px;
-    --consonant-merch-card-twp-mobile-width: 300px;
-    --consonant-merch-card-twp-mobile-height: 358px;
-
-    /* ccd-action */
-    --consonant-merch-card-ccd-action-width: 276px;
-    --consonant-merch-card-ccd-action-min-height: 320px;
-
-
-    /*mini compare chart */
-    --consonant-merch-card-mini-compare-chart-icon-size: 32px;
-    --consonant-merch-card-mini-compare-mobile-cta-font-size: 15px;
-    --consonant-merch-card-mini-compare-mobile-cta-width: 75px;
-    --consonant-merch-card-mini-compare-badge-mobile-max-width: 50px;
 
     /* inline SVGs */
     --checkmark-icon: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 10 10'%3E%3Cpath fill='%23fff' d='M3.788 9A.999.999 0 0 1 3 8.615l-2.288-3a1 1 0 1 1 1.576-1.23l1.5 1.991 3.924-4.991a1 1 0 1 1 1.576 1.23l-4.712 6A.999.999 0 0 1 3.788 9z' class='spectrum-UIIcon--medium'/%3E%3C/svg%3E%0A");
@@ -581,10 +1403,6 @@ merch-card.has-divider hr {
     margin-bottom: var(--consonant-merch-spacing-xs);
     height: 1px;
     border: none;
-}
-
-merch-card[variant="special-offers"] span[is="inline-price"][data-template="strikethrough"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
 }
 
 merch-card p, merch-card h3, merch-card h4 {
@@ -741,153 +1559,10 @@ merch-card [slot="body-xl"] {
     color: var(--merch-color-grey-80);
 }
 
-merch-card[variant="plans"] [slot="description"] {
-    min-height: 84px;
-}
-
-merch-card[variant="catalog"] [slot="action-menu-content"] {
-    background-color: #000;
-    color: var(--color-white, #fff);
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    width: fit-content;
-    padding: var(--consonant-merch-spacing-xs);
-    border-radius: var(--consonant-merch-spacing-xxxs);
-    position: absolute;
-    top: 55px;
-    right: 15px;
-    line-height: var(--consonant-merch-card-body-line-height);
-}
-
-merch-card[variant="catalog"] [slot="action-menu-content"] ul {
-    padding-left: 0;
-    padding-bottom: var(--consonant-merch-spacing-xss);
-    margin-top: 0;
-    margin-bottom: 0;
-    list-style-position: inside;
-    list-style-type: '\u2022 ';
-}
-
-merch-card[variant="catalog"] [slot="action-menu-content"] ul li {
-    padding-left: 0;
-    line-height: var(--consonant-merch-card-body-line-height);
-}
-
-merch-card[variant="catalog"] [slot="action-menu-content"] ::marker {
-    margin-right: 0;
-}
-
-merch-card[variant="catalog"] [slot="action-menu-content"] p {
-    color: var(--color-white, #fff);
-}
-
-merch-card[variant="catalog"] [slot="action-menu-content"] a {
-    color: var(--consonant-merch-card-background-color);
-    text-decoration: underline;
-}
-
-merch-card[variant="catalog"] .payment-details {
-  font-size: var(--consonant-merch-card-body-font-size);
-  font-style: italic;
-  font-weight: 400;
-  line-height: var(--consonant-merch-card-body-line-height);
-}
-
-merch-card[variant="ccd-action"] .price-strikethrough {
-    font-size: 18px;
-}
-
-merch-card[variant="plans"] [slot="quantity-select"] {
-    display: flex;
-    justify-content: flex-start;
-    box-sizing: border-box;
-    width: 100%;
-    padding: var(--consonant-merch-spacing-xs);
-}
-
-merch-card[variant="twp"] div[class$='twp-badge'] {
-    padding: 4px 10px 5px 10px;
-}
-
-merch-card[variant="twp"] [slot="body-xs-top"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-    color: var(--merch-color-grey-80);
-}
-
-merch-card[variant="twp"] [slot="body-xs"] ul {
-    padding: 0;
-    margin: 0;
-}
-
-merch-card[variant="twp"] [slot="body-xs"] ul li {
-    list-style-type: none;
-    padding-left: 0;
-}
-
-merch-card[variant="twp"] [slot="body-xs"] ul li::before {
-    content: '\xB7';
-    font-size: 20px;
-    padding-right: 5px;
-    font-weight: bold;
-}
-
-merch-card[variant="twp"] [slot="footer"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-    padding: var(--consonant-merch-spacing-s)
-    var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-xs);
-    color: var(--merch-color-grey-80);
-    display: flex;
-    flex-flow: wrap;
-}
-
-merch-card[variant='twp'] merch-quantity-select,
-merch-card[variant='twp'] merch-offer-select {
-    display: none;
-}
-
 [slot="cci-footer"] p,
 [slot="cct-footer"] p,
 [slot="cce-footer"] p {
     margin: 0;
-}
-
-/* mini compare chart card styles */
-
-merch-card[variant="mini-compare-chart"] [slot="heading-m"] {
-    padding: 0 var(--consonant-merch-spacing-s) 0;
-}
-
-merch-card[variant="mini-compare-chart"] [slot="body-m"] {
-    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s);
-}
-
-merch-card[variant="mini-compare-chart"] [is="inline-price"] {
-    display: inline-block;
-    min-height: 30px;
-    min-width: 1px;
-}
-
-merch-card[variant="mini-compare-chart"] [slot='callout-content'] {
-    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s) 0px;
-}
-
-merch-card[variant="mini-compare-chart"] [slot='callout-content'] [is="inline-price"] {
-    min-height: unset;
-}
-
-merch-card[variant="mini-compare-chart"] [slot="price-commitment"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    padding: 0 var(--consonant-merch-spacing-s);
-}
-
-merch-card[variant="mini-compare-chart"] [slot="price-commitment"] a {
-    display: inline-block;
-    height: 27px;
-}
-
-merch-card[variant="mini-compare-chart"] [slot="offers"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
 }
 
 merch-card [slot="promo-text"] {
@@ -898,111 +1573,6 @@ merch-card [slot="promo-text"] {
     margin: 0;
     min-height: var(--consonant-merch-card-promo-text-height);
     padding: 0;
-}
-
-
-merch-card[variant="mini-compare-chart"] [slot="body-xxs"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s) 0;    
-}
-
-merch-card[variant="mini-compare-chart"] [slot="promo-text"] {
-    font-size: var(--consonant-merch-card-body-m-font-size);
-    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s) 0;
-}
-
-merch-card[variant="mini-compare-chart"] [slot="promo-text"] a {
-    text-decoration: underline;
-}
-
-merch-card[variant="mini-compare-chart"] .footer-row-icon {
-    display: flex;
-    place-items: center;
-}
-
-merch-card[variant="mini-compare-chart"] .footer-row-icon img {
-    max-width: initial;
-    width: var(--consonant-merch-card-mini-compare-chart-icon-size);
-    height: var(--consonant-merch-card-mini-compare-chart-icon-size);
-}
-
-merch-card[variant="mini-compare-chart"] .footer-row-cell {
-    border-top: 1px solid var(--consonant-merch-card-border-color);
-    display: flex;
-    gap: var(--consonant-merch-spacing-xs);
-    justify-content: start;
-    place-items: center;
-    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s);
-}
-
-merch-card[variant="mini-compare-chart"] .footer-row-cell-description {
-    font-size: var(--consonant-merch-card-body-s-font-size);
-    line-height: var(--consonant-merch-card-body-s-line-height);
-}
-
-merch-card[variant="mini-compare-chart"] .footer-row-cell-description p {
-    color: var(--merch-color-grey-80);
-    vertical-align: bottom;
-}
-
-merch-card[variant="mini-compare-chart"] .footer-row-cell-description a {
-    color: var(--color-accent);
-    text-decoration: solid;
-}
-
-/* mini compare mobile */ 
-@media screen and ${v} {
-  merch-card[variant="mini-compare-chart"] [slot='heading-m'] {
-    font-size: var(--consonant-merch-card-body-s-font-size);
-    line-height: var(--consonant-merch-card-body-s-line-height);
-  }
-
-  merch-card[variant="mini-compare-chart"] [slot='heading-m-price'] {
-    font-size: var(--consonant-merch-card-body-s-font-size);
-    line-height: var(--consonant-merch-card-body-s-line-height);
-  }
-
-  merch-card[variant="mini-compare-chart"] [slot='body-m'] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-  }
-  
-  merch-card[variant="mini-compare-chart"] [slot="promo-text"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-  }
-  merch-card[variant="mini-compare-chart"] .footer-row-cell-description {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-  }
-}
-
-/* mini compare tablet */
-@media screen and ${b} {
-  merch-card[variant="mini-compare-chart"] [slot='heading-m'] {
-    font-size: var(--consonant-merch-card-body-s-font-size);
-    line-height: var(--consonant-merch-card-body-s-line-height);
-  }
-
-  merch-card[variant="mini-compare-chart"] [slot='heading-m-price'] {
-    font-size: var(--consonant-merch-card-body-s-font-size);
-    line-height: var(--consonant-merch-card-body-s-line-height);
-  }
-
-  merch-card[variant="mini-compare-chart"] [slot='body-m'] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-  }
-
-  merch-card[variant="mini-compare-chart"] [slot="promo-text"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-  }
-  
-  merch-card[variant="mini-compare-chart"] .footer-row-cell-description {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-  }
 }
 
 div[slot="footer"] {
@@ -1027,451 +1597,6 @@ div[slot='bg-image'] img {
     object-fit: cover;
     border-top-left-radius: 16px;
     border-top-right-radius: 16px;
-}
-
-/* Mobile */
-@media screen and ${v} {
-    :root {
-        --consonant-merch-card-mini-compare-chart-width: 302px;
-        --consonant-merch-card-segment-width: 276px;
-        --consonant-merch-card-mini-compare-chart-wide-width: 302px;
-        --consonant-merch-card-special-offers-width: 302px;
-        --consonant-merch-card-twp-width: 300px;
-    }
-}
-
-
-/* Tablet */
-@media screen and ${h} {
-    :root {
-        --consonant-merch-card-catalog-width: 302px;
-        --consonant-merch-card-plans-width: 302px;
-        --consonant-merch-card-segment-width: 276px;
-        --consonant-merch-card-mini-compare-chart-width: 302px;
-        --consonant-merch-card-mini-compare-chart-wide-width: 302px;
-        --consonant-merch-card-special-offers-width: 302px;
-        --consonant-merch-card-twp-width: 268px;
-    }
-}
-
-/* desktop */
-@media screen and ${s} {
-    :root {
-        --consonant-merch-card-catalog-width: 276px;
-        --consonant-merch-card-plans-width: 276px;
-        --consonant-merch-card-segment-width: 302px;
-        --consonant-merch-card-inline-heading-width: 378px;
-        --consonant-merch-card-product-width: 378px;
-        --consonant-merch-card-image-width: 378px;
-        --consonant-merch-card-mini-compare-chart-width: 378px;
-        --consonant-merch-card-mini-compare-chart-wide-width: 484px;
-        --consonant-merch-card-twp-width: 268px;
-    }
-}
-
-/* supported cards */
-/* grid style for plans */
-.one-merch-card.plans,
-.two-merch-cards.plans,
-.three-merch-cards.plans,
-.four-merch-cards.plans {
-    grid-template-columns: var(--consonant-merch-card-plans-width);
-}
-
-/* Tablet */
-@media screen and ${h} {
-    .two-merch-cards.plans,
-    .three-merch-cards.plans,
-    .four-merch-cards.plans {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-plans-width));
-    }
-}
-
-/* desktop */
-@media screen and ${s} {
-    .three-merch-cards.plans,
-    .four-merch-cards.plans {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-plans-width));
-    }
-}
-
-/* Large desktop */
-    @media screen and ${d} {
-    .four-merch-cards.plans {
-        grid-template-columns: repeat(4, var(--consonant-merch-card-plans-width));
-    }
-}
-
-
-/* grid style for catalog */
-.one-merch-card.catalog,
-.two-merch-cards.catalog,
-.three-merch-cards.catalog,
-.four-merch-cards.catalog {
-    grid-template-columns: var(--consonant-merch-card-catalog-width);
-}
-
-/* Tablet */
-@media screen and ${h} {
-    .two-merch-cards.catalog,
-    .three-merch-cards.catalog,
-    .four-merch-cards.catalog {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-catalog-width));
-    }
-}
-
-/* desktop */
-@media screen and ${s} {
-    .three-merch-cards.catalog,
-    .four-merch-cards.catalog {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-catalog-width));
-    }
-}
-
-/* Large desktop */
-    @media screen and ${d} {
-    .four-merch-cards.catalog {
-        grid-template-columns: repeat(4, var(--consonant-merch-card-catalog-width));
-    }
-}
-
-
-/* grid style for special-offers */
-.one-merch-card.special-offers,
-.two-merch-cards.special-offers,
-.three-merch-cards.special-offers,
-.four-merch-cards.special-offers {
-    grid-template-columns: minmax(300px, var(--consonant-merch-card-special-offers-width));
-}
-
-/* Tablet */
-@media screen and ${h} {
-    .two-merch-cards.special-offers,
-    .three-merch-cards.special-offers,
-    .four-merch-cards.special-offers {
-        grid-template-columns: repeat(2, minmax(300px, var(--consonant-merch-card-special-offers-width)));
-    }
-}
-
-/* desktop */
-@media screen and ${s} {
-    .three-merch-cards.special-offers,
-    .four-merch-cards.special-offers {
-        grid-template-columns: repeat(3, minmax(300px, var(--consonant-merch-card-special-offers-width)));
-    }
-}
-
-@media screen and ${d} {
-    .four-merch-cards.special-offers {
-        grid-template-columns: repeat(4, minmax(300px, var(--consonant-merch-card-special-offers-width)));
-    }
-}
-
-
-/* grid style for image */
-.one-merch-card.image,
-.two-merch-cards.image,
-.three-merch-cards.image,
-.four-merch-cards.image {
-    grid-template-columns: var(--consonant-merch-card-image-width);
-}
-
-/* Tablet */
-@media screen and ${h} {
-    .two-merch-cards.image,
-    .three-merch-cards.image,
-    .four-merch-cards.image {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-image-width));
-    }
-}
-
-/* desktop */
-@media screen and ${s} {
-    .three-merch-cards.image,
-    .four-merch-cards.image {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-image-width));
-    }
-}
-
-/* Large desktop */
-    @media screen and ${d} {
-    .four-merch-cards.image {
-        grid-template-columns: repeat(4, var(--consonant-merch-card-image-width));
-    }
-}
-
-
-/* grid style for segment */
-.one-merch-card.segment,
-.two-merch-cards.segment,
-.three-merch-cards.segment,
-.four-merch-cards.segment {
-    grid-template-columns: minmax(276px, var(--consonant-merch-card-segment-width));
-}
-
-/* Tablet */
-@media screen and ${h} {
-    .two-merch-cards.segment,
-    .three-merch-cards.segment,
-    .four-merch-cards.segment {
-        grid-template-columns: repeat(2, minmax(276px, var(--consonant-merch-card-segment-width)));
-    }
-}
-
-/* desktop */
-@media screen and ${s} {
-    .three-merch-cards.segment {
-        grid-template-columns: repeat(3, minmax(276px, var(--consonant-merch-card-segment-width)));
-    }
-
-    .four-merch-cards.segment {
-        grid-template-columns: repeat(4, minmax(276px, var(--consonant-merch-card-segment-width)));
-    }
-}
-
-
-/* grid style for product */
-.one-merch-card.product,
-.two-merch-cards.product,
-.three-merch-cards.product,
-.four-merch-cards.product {
-    grid-template-columns: var(--consonant-merch-card-product-width);
-}
-
-/* Tablet */
-@media screen and ${h} {
-    .two-merch-cards.product,
-    .three-merch-cards.product,
-    .four-merch-cards.product {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-product-width));
-    }
-}
-
-/* desktop */
-@media screen and ${s} {
-    .three-merch-cards.product,
-    .four-merch-cards.product {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-product-width));
-    }
-}
-
-/* Large desktop */
-    @media screen and ${d} {
-    .four-merch-cards.product {
-        grid-template-columns: repeat(4, var(--consonant-merch-card-product-width));
-    }
-}
-
-/* grid style for twp */
-.one-merch-card.twp,
-.two-merch-cards.twp,
-.three-merch-cards.twp {
-    grid-template-columns: var(--consonant-merch-card-image-width);
-}
-
-/* Tablet */
-@media screen and ${h} {
-    .one-merch-card.twp,
-    .two-merch-cards.twp {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-twp-width));
-    }
-    .three-merch-cards.twp {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-twp-width));
-    }
-}
-
-/* desktop */
-@media screen and ${s} {
-    .one-merch-card.twp
-    .two-merch-cards.twp {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-twp-width));
-    }
-    .three-merch-cards.twp {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-twp-width));
-    }
-}
-
-/* Large desktop */
-    @media screen and ${d} {
-        .one-merch-card.twp
-        .two-merch-cards.twp {
-            grid-template-columns: repeat(2, var(--consonant-merch-card-twp-width));
-        }
-        .three-merch-cards.twp {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-twp-width));
-    }
-}
-
-/* Mobile */
-@media screen and ${v} {
-    .one-merch-card.twp,
-    .two-merch-cards.twp,
-    .three-merch-cards.twp {
-        grid-template-columns: repeat(1, var(--consonant-merch-card-twp-mobile-width));
-    }
-}
-
-/* grid style for inline-heading */
-.one-merch-card.inline-heading,
-.two-merch-cards.inline-heading,
-.three-merch-cards.inline-heading,
-.four-merch-cards.inline-heading {
-    grid-template-columns: var(--consonant-merch-card-inline-heading-width);
-}
-
-/* Tablet */
-@media screen and ${h} {
-    .two-merch-cards.inline-heading,
-    .three-merch-cards.inline-heading,
-    .four-merch-cards.inline-heading {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-inline-heading-width));
-    }
-}
-
-/* desktop */
-@media screen and ${s} {
-    .three-merch-cards.inline-heading,
-    .four-merch-cards.inline-heading {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-inline-heading-width));
-    }
-}
-
-/* Large desktop */
-    @media screen and ${d} {
-    .four-merch-cards.inline-heading {
-        grid-template-columns: repeat(4, var(--consonant-merch-card-inline-heading-width));
-    }
-}
-
-/* grid style for ccd-action */
-.one-merch-card.ccd-action,
-.two-merch-cards.ccd-action,
-.three-merch-cards.ccd-action,
-.four-merch-cards.ccd-action {
-    grid-template-columns: var(--consonant-merch-card-ccd-action-width);
-}
-
-/* Tablet */
-@media screen and ${h} {
-    .two-merch-cards.ccd-action,
-    .three-merch-cards.ccd-action,
-    .four-merch-cards.ccd-action {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-ccd-action-width));
-    }
-}
-
-/* desktop */
-@media screen and ${s} {
-    .three-merch-cards.ccd-action,
-    .four-merch-cards.ccd-action {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-ccd-action-width));
-    }
-}
-
-/* Large desktop */
-    @media screen and ${d} {
-    .four-merch-cards.ccd-action {
-        grid-template-columns: repeat(4, var(--consonant-merch-card-ccd-action-width));
-    }
-}
-
-/* grid style for mini-compare-chart */
-.one-merch-card.mini-compare-chart {
-    grid-template-columns: var(--consonant-merch-card-mini-compare-chart-wide-width);
-    gap: var(--consonant-merch-spacing-xs);
-}
-
-.two-merch-cards.mini-compare-chart,
-.three-merch-cards.mini-compare-chart,
-.four-merch-cards.mini-compare-chart {
-    grid-template-columns: repeat(2, var(--consonant-merch-card-mini-compare-chart-width));
-    gap: var(--consonant-merch-spacing-xs);
-}
-
-@media screen and ${v} {
-    .two-merch-cards.mini-compare-chart,
-    .three-merch-cards.mini-compare-chart,
-    .four-merch-cards.mini-compare-chart {
-        grid-template-columns: var(--consonant-merch-card-mini-compare-chart-width);
-        gap: var(--consonant-merch-spacing-xs);
-    }
-}
-
-@media screen and ${b} {
-    .three-merch-cards.mini-compare-chart merch-card [slot="footer"] a,
-    .four-merch-cards.mini-compare-chart merch-card [slot="footer"] a {
-        flex: 1;
-    }
-}
-
-/* Tablet */
-@media screen and ${h} {
-    .two-merch-cards.mini-compare-chart {
-        grid-template-columns: repeat(2, minmax(var(--consonant-merch-card-mini-compare-chart-width), var(--consonant-merch-card-mini-compare-chart-wide-width)));
-        gap: var(--consonant-merch-spacing-m);
-    }
-
-    .three-merch-cards.mini-compare-chart,
-    .four-merch-cards.mini-compare-chart {
-        grid-template-columns: repeat(2, minmax(var(--consonant-merch-card-mini-compare-chart-width), var(--consonant-merch-card-mini-compare-chart-wide-width)));
-    }
-}
-
-/* desktop */
-@media screen and ${s} {
-    .one-merch-card.mini-compare-chart {
-        grid-template-columns: var(--consonant-merch-card-mini-compare-chart-wide-width);
-    }
-
-    .two-merch-cards.mini-compare-chart {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-mini-compare-chart-wide-width));
-        gap: var(--consonant-merch-spacing-m);
-    }
-
-    .three-merch-cards.mini-compare-chart,
-    .four-merch-cards.mini-compare-chart {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-mini-compare-chart-width));
-        gap: var(--consonant-merch-spacing-m);
-    }
-}
-
-@media screen and ${d} {
-    .four-merch-cards.mini-compare-chart {
-        grid-template-columns: repeat(4, var(--consonant-merch-card-mini-compare-chart-width));
-    }
-}
-
-/* mini-compare card footer rows */
-merch-card .footer-row-cell:nth-child(1) {
-    min-height: var(--consonant-merch-card-footer-row-1-min-height);
-}
-
-merch-card .footer-row-cell:nth-child(2) {
-    min-height: var(--consonant-merch-card-footer-row-2-min-height);
-}
-
-merch-card .footer-row-cell:nth-child(3) {
-    min-height: var(--consonant-merch-card-footer-row-3-min-height);
-}
-
-merch-card .footer-row-cell:nth-child(4) {
-    min-height: var(--consonant-merch-card-footer-row-4-min-height);
-}
-
-merch-card .footer-row-cell:nth-child(5) {
-    min-height: var(--consonant-merch-card-footer-row-5-min-height);
-}
-
-merch-card .footer-row-cell:nth-child(6) {
-    min-height: var(--consonant-merch-card-footer-row-6-min-height);
-}
-
-merch-card .footer-row-cell:nth-child(7) {
-    min-height: var(--consonant-merch-card-footer-row-7-min-height);
-}
-
-merch-card .footer-row-cell:nth-child(8) {
-    min-height: var(--consonant-merch-card-footer-row-8-min-height);
 }
 
 span[is="inline-price"][data-template='strikethrough'] {
@@ -1499,152 +1624,5 @@ body.merch-modal {
     scrollbar-gutter: stable;
     height: 100vh;
 }
-`;document.head.appendChild(q);var g="merch-card",Z=32,E="mini-compare-chart",j=c=>`--consonant-merch-card-footer-row-${c}-min-height`,p=class extends K{static properties={name:{type:String,attribute:"name",reflect:!0},variant:{type:String,reflect:!0},size:{type:String,attribute:"size",reflect:!0},badgeColor:{type:String,attribute:"badge-color"},borderColor:{type:String,attribute:"border-color"},badgeBackgroundColor:{type:String,attribute:"badge-background-color"},badgeText:{type:String,attribute:"badge-text"},actionMenu:{type:Boolean,attribute:"action-menu"},actionMenuContent:{type:String,attribute:"action-menu-content"},customHr:{type:Boolean,attribute:"custom-hr"},detailBg:{type:String,attribute:"detail-bg"},secureLabel:{type:String,attribute:"secure-label"},checkboxLabel:{type:String,attribute:"checkbox-label"},selected:{type:Boolean,attribute:"aria-selected",reflect:!0},storageOption:{type:String,attribute:"storage",reflect:!0},stockOfferOsis:{type:Object,attribute:"stock-offer-osis",converter:{fromAttribute:e=>{let[t,r,n]=e.split(",");return{PUF:t,ABM:r,M2M:n}}}},filters:{type:String,reflect:!0,converter:{fromAttribute:e=>Object.fromEntries(e.split(",").map(t=>{let[r,n,a]=t.split(":"),i=Number(n);return[r,{order:isNaN(i)?void 0:i,size:a}]})),toAttribute:e=>Object.entries(e).map(([t,{order:r,size:n}])=>[t,r,n].filter(a=>a!=null).join(":")).join(",")}},types:{type:String,attribute:"types",reflect:!0},merchOffer:{type:Object}};static styles=[D,...I()];customerSegment;marketSegment;constructor(){super(),this.filters={},this.types="",this.selected=!1}#e;updated(e){(e.has("badgeBackgroundColor")||e.has("borderColor"))&&(this.style.border=this.computedBorderStyle),this.updateComplete.then(async()=>{let r=Array.from(this.querySelectorAll('span[is="inline-price"][data-wcs-osi]')).filter(n=>!n.closest('[slot="callout-content"]'));await Promise.all(r.map(n=>n.onceSettled())),this.adjustTitleWidth(),O()?this.removeEmptyRows():(this.adjustMiniCompareBodySlots(),this.adjustMiniCompareFooterRows())})}get computedBorderStyle(){return this.variant!=="twp"?`1px solid ${this.borderColor?this.borderColor:this.badgeBackgroundColor}`:""}get evergreen(){return this.classList.contains("intro-pricing")}get stockCheckbox(){return this.checkboxLabel?o`<label id="stock-checkbox">
-                    <input type="checkbox" @change=${this.toggleStockOffer}></input>
-                    <span></span>
-                    ${this.checkboxLabel}
-                </label>`:""}get cardImage(){return o` <div class="image">
-            <slot name="bg-image"></slot>
-            ${this.badge}
-        </div>`}get secureLabelFooter(){let e=this.secureLabel?o`<span class="secure-transaction-label"
-                  >${this.secureLabel}</span
-              >`:"";return o`<footer>${e}<slot name="footer"></slot></footer>`}get miniCompareFooter(){let e=this.secureLabel?o`<slot name="secure-transaction-label">
-                  <span class="secure-transaction-label"
-                      >${this.secureLabel}</span
-                  ></slot
-              >`:o`<slot name="secure-transaction-label"></slot>`;return o`<footer>${e}<slot name="footer"></slot></footer>`}get badge(){let e;if(!(!this.badgeBackgroundColor||!this.badgeColor||!this.badgeText))return this.evergreen&&(e=`border: 1px solid ${this.badgeBackgroundColor}; border-right: none;`),o`
-            <div
-                id="badge"
-                class="${this.variant}-badge"
-                style="background-color: ${this.badgeBackgroundColor};
-                    color: ${this.badgeColor};
-                    ${e}"
-            >
-                ${this.badgeText}
-            </div>
-        `}get badgeElement(){return this.shadowRoot.getElementById("badge")}getContainer(){return this.closest('[class*="-merch-cards"]')??this.parentElement}get headingmMSlot(){return this.shadowRoot.querySelector('slot[name="heading-m"]').assignedElements()[0]}get footerSlot(){return this.shadowRoot.querySelector('slot[name="footer"]')?.assignedElements()[0]}get price(){return this.headingmMSlot?.querySelector('span[is="inline-price"]')}get checkoutLinks(){return[...this.footerSlot?.querySelectorAll('a[is="checkout-link"]')??[]]}async toggleStockOffer({target:e}){if(!this.stockOfferOsis)return;let t=this.checkoutLinks;if(t.length!==0)for(let r of t){await r.onceSettled();let n=r.value?.[0]?.planType;if(!n)return;let a=this.stockOfferOsis[n];if(!a)return;let i=r.dataset.wcsOsi.split(",").filter(u=>u!==a);e.checked&&i.push(a),r.dataset.wcsOsi=i.join(",")}}toggleActionMenu(e){let t=e?.type==="mouseleave"?!0:void 0,r=this.shadowRoot.querySelector('slot[name="action-menu-content"]');r&&(t||this.dispatchEvent(new CustomEvent(R,{bubbles:!0,composed:!0,detail:{card:this.name,type:"action-menu"}})),r.classList.toggle("hidden",t))}handleQuantitySelection(e){let t=this.checkoutLinks;for(let r of t)r.dataset.quantity=e.detail.option}get titleElement(){return this.variant==="special-offers"?this.querySelector('[slot="detail-m"]'):this.querySelector('[slot="heading-xs"]')}get title(){return this.titleElement?.textContent?.trim()}get description(){return this.querySelector('[slot="body-xs"]')?.textContent?.trim()}updateFilters(e){let t={...this.filters};Object.keys(t).forEach(r=>{if(e){t[r].order=Math.min(t[r].order||2,2);return}let n=t[r].order;n===1||isNaN(n)||(t[r].order=Number(n)+1)}),this.filters=t}includes(e){return this.textContent.match(new RegExp(e,"i"))!==null}render(){if(!(!this.isConnected||this.style.display==="none"))switch(this.variant){case"special-offers":return this.renderSpecialOffer();case"segment":return this.renderSegment();case"plans":return this.renderPlans();case"catalog":return this.renderCatalog();case"image":return this.renderImage();case"product":return this.renderProduct();case"inline-heading":return this.renderInlineHeading();case E:return this.renderMiniCompareChart();case"ccd-action":return this.renderCcdAction();case"twp":return this.renderTwp();default:return this.renderProduct()}}renderSpecialOffer(){return o`${this.cardImage}
-            <div class="body">
-                <slot name="detail-m"></slot>
-                <slot name="heading-xs"></slot>
-                <slot name="body-xs"></slot>
-            </div>
-            ${this.evergreen?o`
-                      <div
-                          class="detail-bg-container"
-                          style="background: ${this.detailBg}"
-                      >
-                          <slot name="detail-bg"></slot>
-                      </div>
-                  `:o`
-                      <hr />
-                      ${this.secureLabelFooter}
-                  `}`}get promoBottom(){return this.classList.contains("promo-bottom")}get startingAt(){return this.classList.contains("starting-at")}renderSegment(){return o` ${this.badge}
-            <div class="body">
-                <slot name="heading-xs"></slot>
-                <slot name="body-xxs"></slot>
-                ${this.promoBottom?"":o`<slot name="promo-text"></slot><slot name="callout-content"></slot>`}
-                <slot name="body-xs"></slot>
-                ${this.promoBottom?o`<slot name="promo-text"></slot><slot name="callout-content"></slot>`:""}
-            </div>
-            <hr />
-            ${this.secureLabelFooter}`}renderPlans(){return o` ${this.badge}
-            <div class="body">
-                <slot name="icons"></slot>
-                <slot name="heading-xs"></slot>
-                <slot name="heading-m"></slot>
-                <slot name="body-xxs"></slot>
-                ${this.promoBottom?"":o`<slot name="promo-text"></slot><slot name="callout-content"></slot> `}
-                <slot name="body-xs"></slot>
-                ${this.promoBottom?o`<slot name="promo-text"></slot><slot name="callout-content"></slot> `:""}  
-                ${this.stockCheckbox}
-            </div>
-            <slot name="quantity-select"></slot>
-            ${this.secureLabelFooter}`}renderCatalog(){return o` <div class="body">
-                <div class="top-section">
-                    <slot name="icons"></slot> ${this.badge}
-                    <div
-                        class="action-menu
-                        ${N()&&this.actionMenu?"always-visible":""}
-                        ${this.actionMenu?"invisible":"hidden"}"
-                        @click="${this.toggleActionMenu}"
-                    ></div>
-                </div>
-                <slot
-                    name="action-menu-content"
-                    class="action-menu-content
-                    ${this.actionMenuContent?"":"hidden"}"
-                    >${this.actionMenuContent}</slot
-                >
-                <slot name="heading-xs"></slot>
-                <slot name="heading-m"></slot>
-                <slot name="body-xxs"></slot>
-                ${this.promoBottom?"":o`<slot name="promo-text"></slot
-                          ><slot name="callout-content"></slot>`}
-                <slot name="body-xs"></slot>
-                ${this.promoBottom?o`<slot name="promo-text"></slot
-                          ><slot name="callout-content"></slot>`:""}
-            </div>
-            ${this.secureLabelFooter}`}renderImage(){return o`${this.cardImage}
-            <div class="body">
-                <slot name="icons"></slot>
-                <slot name="heading-xs"></slot>
-                <slot name="body-xxs"></slot>
-                ${this.promoBottom?o`<slot name="body-xs"></slot><slot name="promo-text"></slot>`:o`<slot name="promo-text"></slot><slot name="body-xs"></slot>`}
-            </div>
-            ${this.evergreen?o`
-                      <div
-                          class="detail-bg-container"
-                          style="background: ${this.detailBg}"
-                      >
-                          <slot name="detail-bg"></slot>
-                      </div>
-                  `:o`
-                      <hr />
-                      ${this.secureLabelFooter}
-                  `}`}renderInlineHeading(){return o` ${this.badge}
-            <div class="body">
-                <div class="top-section">
-                    <slot name="icons"></slot>
-                    <slot name="heading-xs"></slot>
-                </div>
-                <slot name="body-xs"></slot>
-            </div>
-            ${this.customHr?"":o`<hr />`} ${this.secureLabelFooter}`}renderProduct(){return o` ${this.badge}
-            <div class="body">
-                <slot name="icons"></slot>
-                <slot name="heading-xs"></slot>
-                <slot name="body-xxs"></slot>
-                ${this.promoBottom?"":o`<slot name="promo-text"></slot><slot name="callout-content"></slot>`}
-                <slot name="body-xs"></slot>
-                ${this.promoBottom?o`<slot name="promo-text"></slot><slot name="callout-content"></slot>`:""}
-            </div>
-            ${this.secureLabelFooter}`}renderMiniCompareChart(){let{badge:e}=this;return o` <div class="top-section${e?" badge":""}">
-                <slot name="icons"></slot> ${e}
-            </div>
-            <slot name="heading-m"></slot>
-            <slot name="body-m"></slot>
-            <slot name="heading-m-price"></slot>
-            <slot name="body-xxs"></slot>
-            <slot name="price-commitment"></slot>
-            <slot name="offers"></slot>
-            <slot name="promo-text"></slot>
-            <slot name="callout-content"></slot>
-            ${this.miniCompareFooter}
-            <slot name="footer-rows"><slot name="body-s"></slot></slot>`}renderTwp(){return o`${this.badge}
-            <div class="top-section">
-                <slot name="icons"></slot>
-                <slot name="heading-xs"></slot>
-                <slot name="body-xs-top"></slot>
-            </div>
-            <div class="body">
-                <slot name="body-xs"></slot>
-            </div>
-            <footer><slot name="footer"></slot></footer>`}get defaultSlot(){return this.querySelector(":scope > a:not([slot]),:scope > p:not([slot]),:scope > div:not([slot]),:scope > span:not([slot])")?o`<slot></slot>`:Q}renderCcdAction(){return o` <div class="body">
-            <slot name="icons"></slot> ${this.badge}
-            <slot name="heading-xs"></slot>
-            <slot name="heading-m"></slot>
-            ${this.promoBottom?o`<slot name="body-xs"></slot><slot name="promo-text"></slot>`:o`<slot name="promo-text"></slot><slot name="body-xs"></slot>`}
-            <footer><slot name="footer"></slot></footer>
-            ${this.defaultSlot}
-        </div>`}connectedCallback(){super.connectedCallback(),this.#e=this.getContainer(),this.setAttribute("tabindex",this.getAttribute("tabindex")??"0"),this.addEventListener("mouseleave",this.toggleActionMenu),this.addEventListener(k,this.handleQuantitySelection),this.addEventListener(_,this.merchCardReady,{once:!0}),this.updateComplete.then(()=>{this.merchCardReady()}),this.storageOptions?.addEventListener("change",this.handleStorageChange)}disconnectedCallback(){super.disconnectedCallback(),this.removeEventListener(k,this.handleQuantitySelection),this.storageOptions?.removeEventListener(C,this.handleStorageChange)}appendInvisibleSpacesToFooterLinks(){[...this.querySelectorAll('[slot="footer"] a')].forEach(e=>{L(e).forEach(r=>{let i=r.textContent.split(" ").map(u=>u.match(/.{1,7}/g)?.join("\u200B")).join(" ");r.textContent=i})})}updateMiniCompareElementMinHeight(e,t){let r=`--consonant-merch-card-mini-compare-${t}-height`,n=Math.max(0,parseInt(window.getComputedStyle(e).height)||0),a=parseInt(this.#e.style.getPropertyValue(r))||0;n>a&&this.#e.style.setProperty(r,`${n}px`)}async adjustTitleWidth(){if(!["segment","plans"].includes(this.variant))return;let e=this.getBoundingClientRect().width,t=this.badgeElement?.getBoundingClientRect().width||0;e===0||t===0||this.style.setProperty("--consonant-merch-card-heading-xs-max-width",`${Math.round(e-t-16)}px`)}async adjustMiniCompareBodySlots(){if(this.variant!==E||this.getBoundingClientRect().width===0)return;this.updateMiniCompareElementMinHeight(this.shadowRoot.querySelector(".top-section"),"top-section"),["heading-m","body-m","heading-m-price","price-commitment","offers","promo-text","callout-content","secure-transaction-label"].forEach(r=>this.updateMiniCompareElementMinHeight(this.shadowRoot.querySelector(`slot[name="${r}"]`),r)),this.updateMiniCompareElementMinHeight(this.shadowRoot.querySelector("footer"),"footer");let t=this.shadowRoot.querySelector(".mini-compare-chart-badge");t&&t.textContent!==""&&this.#e.style.setProperty("--consonant-merch-card-mini-compare-top-section-mobile-height","32px")}adjustMiniCompareFooterRows(){if(this.variant!==E||this.getBoundingClientRect().width===0)return;[...this.querySelector('[slot="footer-rows"]').children].forEach((t,r)=>{let n=Math.max(Z,parseInt(window.getComputedStyle(t).height)||0),a=parseInt(this.#e.style.getPropertyValue(j(r+1)))||0;n>a&&this.#e.style.setProperty(j(r+1),`${n}px`)})}removeEmptyRows(){if(this.variant!==E)return;this.querySelectorAll(".footer-row-cell").forEach(t=>{let r=t.querySelector(".footer-row-cell-description");r&&!r.textContent.trim()&&t.remove()})}get storageOptions(){return this.querySelector("sp-radio-group#storage")}get storageSpecificOfferSelect(){let e=this.storageOptions?.selected;if(e){let t=this.querySelector(`merch-offer-select[storage="${e}"]`);if(t)return t}return this.querySelector("merch-offer-select")}get offerSelect(){return this.storageOptions?this.storageSpecificOfferSelect:this.querySelector("merch-offer-select")}get quantitySelect(){return this.querySelector("merch-quantity-select")}merchCardReady(){this.offerSelect&&!this.offerSelect.planType||this.dispatchEvent(new CustomEvent(M,{bubbles:!0}))}handleStorageChange(){let e=this.closest("merch-card")?.offerSelect.cloneNode(!0);e&&this.dispatchEvent(new CustomEvent(C,{detail:{offerSelect:e},bubbles:!0}))}get dynamicPrice(){return this.querySelector('[slot="price"]')}selectMerchOffer(e){if(e===this.merchOffer)return;this.merchOffer=e;let t=this.dynamicPrice;if(e.price&&t){let r=e.price.cloneNode(!0);t.onceSettled?t.onceSettled().then(()=>{t.replaceWith(r)}):t.replaceWith(r)}}};customElements.define(g,p);
+`;document.head.appendChild(bt);var s="merch-card",i=class extends Nt{static properties={name:{type:String,attribute:"name",reflect:!0},variant:{type:String,reflect:!0},size:{type:String,attribute:"size",reflect:!0},badgeColor:{type:String,attribute:"badge-color"},borderColor:{type:String,attribute:"border-color"},badgeBackgroundColor:{type:String,attribute:"badge-background-color"},badgeText:{type:String,attribute:"badge-text"},actionMenu:{type:Boolean,attribute:"action-menu"},actionMenuContent:{type:String,attribute:"action-menu-content"},customHr:{type:Boolean,attribute:"custom-hr"},detailBg:{type:String,attribute:"detail-bg"},secureLabel:{type:String,attribute:"secure-label"},checkboxLabel:{type:String,attribute:"checkbox-label"},selected:{type:Boolean,attribute:"aria-selected",reflect:!0},storageOption:{type:String,attribute:"storage",reflect:!0},stockOfferOsis:{type:Object,attribute:"stock-offer-osis",converter:{fromAttribute:t=>{let[e,r,n]=t.split(",");return{PUF:e,ABM:r,M2M:n}}}},filters:{type:String,reflect:!0,converter:{fromAttribute:t=>Object.fromEntries(t.split(",").map(e=>{let[r,n,d]=e.split(":"),u=Number(n);return[r,{order:isNaN(u)?void 0:u,size:d}]})),toAttribute:t=>Object.entries(t).map(([e,{order:r,size:n}])=>[e,r,n].filter(d=>d!=null).join(":")).join(",")}},types:{type:String,attribute:"types",reflect:!0},merchOffer:{type:Object}};static styles=[ot,vt(),...nt()];customerSegment;marketSegment;variantLayout;constructor(){super(),this.filters={},this.types="",this.selected=!1}updated(t){(t.has("badgeBackgroundColor")||t.has("borderColor"))&&(this.style.border=this.computedBorderStyle),this.updateComplete.then(async()=>{let r=Array.from(this.querySelectorAll('span[is="inline-price"][data-wcs-osi]')).filter(n=>!n.closest('[slot="callout-content"]'));await Promise.all(r.map(n=>n.onceSettled())),this.variantLayout.postCardUpdateHook(this)})}render(){if(!(!this.isConnected||this.style.display==="none"))return this.variantLayout.renderLayout()}get computedBorderStyle(){return this.variant!=="twp"?`1px solid ${this.borderColor?this.borderColor:this.badgeBackgroundColor}`:""}get badgeElement(){return this.shadowRoot.getElementById("badge")}get headingmMSlot(){return this.shadowRoot.querySelector('slot[name="heading-m"]').assignedElements()[0]}get footerSlot(){return this.shadowRoot.querySelector('slot[name="footer"]')?.assignedElements()[0]}get price(){return this.headingmMSlot?.querySelector('span[is="inline-price"]')}get checkoutLinks(){return[...this.footerSlot?.querySelectorAll('a[is="checkout-link"]')??[]]}async toggleStockOffer({target:t}){if(!this.stockOfferOsis)return;let e=this.checkoutLinks;if(e.length!==0)for(let r of e){await r.onceSettled();let n=r.value?.[0]?.planType;if(!n)return;let d=this.stockOfferOsis[n];if(!d)return;let u=r.dataset.wcsOsi.split(",").filter(v=>v!==d);t.checked&&u.push(d),r.dataset.wcsOsi=u.join(",")}}handleQuantitySelection(t){let e=this.checkoutLinks;for(let r of e)r.dataset.quantity=t.detail.option}get titleElement(){return this.querySelector(this.variantLayout.titleSelector)}get title(){return this.titleElement?.textContent?.trim()}get description(){return this.querySelector('[slot="body-xs"]')?.textContent?.trim()}updateFilters(t){let e={...this.filters};Object.keys(e).forEach(r=>{if(t){e[r].order=Math.min(e[r].order||2,2);return}let n=e[r].order;n===1||isNaN(n)||(e[r].order=Number(n)+1)}),this.filters=e}includes(t){return this.textContent.match(new RegExp(t,"i"))!==null}get startingAt(){return this.classList.contains("starting-at")}get defaultSlot(){return this.querySelector(":scope > a:not([slot]),:scope > p:not([slot]),:scope > div:not([slot]),:scope > span:not([slot])")?Pt`<slot></slot>`:Dt}connectedCallback(){super.connectedCallback(),this.variantLayout=xt(this),this.variantLayout.connectedCallbackHook(),this.setAttribute("tabindex",this.getAttribute("tabindex")??"0"),this.addEventListener(N,this.handleQuantitySelection),this.addEventListener(W,this.merchCardReady,{once:!0}),this.updateComplete.then(()=>{this.merchCardReady()}),this.storageOptions?.addEventListener("change",this.handleStorageChange)}disconnectedCallback(){super.disconnectedCallback(),this.variantLayout.disconnectedCallbackHook(),this.removeEventListener(N,this.handleQuantitySelection),this.storageOptions?.removeEventListener(P,this.handleStorageChange)}appendInvisibleSpacesToFooterLinks(){[...this.querySelectorAll('[slot="footer"] a')].forEach(t=>{Z(t).forEach(r=>{let u=r.textContent.split(" ").map(v=>v.match(/.{1,7}/g)?.join("\u200B")).join(" ");r.textContent=u})})}get storageOptions(){return this.querySelector("sp-radio-group#storage")}get storageSpecificOfferSelect(){let t=this.storageOptions?.selected;if(t){let e=this.querySelector(`merch-offer-select[storage="${t}"]`);if(e)return e}return this.querySelector("merch-offer-select")}get offerSelect(){return this.storageOptions?this.storageSpecificOfferSelect:this.querySelector("merch-offer-select")}get quantitySelect(){return this.querySelector("merch-quantity-select")}merchCardReady(){this.offerSelect&&!this.offerSelect.planType||this.dispatchEvent(new CustomEvent(Y,{bubbles:!0}))}handleStorageChange(){let t=this.closest("merch-card")?.offerSelect.cloneNode(!0);t&&this.dispatchEvent(new CustomEvent(P,{detail:{offerSelect:t},bubbles:!0}))}get dynamicPrice(){return this.querySelector('[slot="price"]')}selectMerchOffer(t){if(t===this.merchOffer)return;this.merchOffer=t;let e=this.dynamicPrice;if(t.price&&e){let r=t.price.cloneNode(!0);e.onceSettled?e.onceSettled().then(()=>{e.replaceWith(r)}):e.replaceWith(r)}}};customElements.define(s,i);
 //# sourceMappingURL=merch-card.js.map
