+++ libs/features/mas/web-components/src/global.css.js
@@ -1,11 +1,3 @@
-import {
-    TABLET_UP,
-    DESKTOP_UP,
-    LARGE_DESKTOP,
-    MOBILE_LANDSCAPE,
-    TABLET_DOWN,
-} from './media.js';
-
 const styles = document.createElement('style');
 
 styles.innerHTML = `
@@ -71,7 +63,6 @@ styles.innerHTML = `
 
 
     --consonant-merch-card-heading-padding: 0;
-    --consonant-merch-card-image-height: 180px;
 
     /* colors */
     --consonant-merch-card-border-color: #eaeaea;
@@ -83,45 +74,6 @@ styles.innerHTML = `
     /* merch card generic */
     --consonant-merch-card-max-width: 300px;
     --transition: cmax-height 0.3s linear, opacity 0.3s linear;
-
-    /* special offers */
-    --consonant-merch-card-special-offers-width: 378px;
-
-    /* image */
-    --consonant-merch-card-image-width: 300px;
-
-    /* segment */
-    --consonant-merch-card-segment-width: 378px;
-
-    /* inline-heading */
-    --consonant-merch-card-inline-heading-width: 300px;
-
-    /* product */
-    --consonant-merch-card-product-width: 300px;
-
-    /* plans */
-    --consonant-merch-card-plans-width: 300px;
-    --consonant-merch-card-plans-icon-size: 40px;
-
-    /* catalog */
-    --consonant-merch-card-catalog-width: 276px;
-    --consonant-merch-card-catalog-icon-size: 40px;
-
-    /* twp */
-    --consonant-merch-card-twp-width: 268px;
-    --consonant-merch-card-twp-mobile-width: 300px;
-    --consonant-merch-card-twp-mobile-height: 358px;
-
-    /* ccd-action */
-    --consonant-merch-card-ccd-action-width: 276px;
-    --consonant-merch-card-ccd-action-min-height: 320px;
-
-
-    /*mini compare chart */
-    --consonant-merch-card-mini-compare-chart-icon-size: 32px;
-    --consonant-merch-card-mini-compare-mobile-cta-font-size: 15px;
-    --consonant-merch-card-mini-compare-mobile-cta-width: 75px;
-    --consonant-merch-card-mini-compare-badge-mobile-max-width: 50px;
 
     /* inline SVGs */
     --checkmark-icon: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 10 10'%3E%3Cpath fill='%23fff' d='M3.788 9A.999.999 0 0 1 3 8.615l-2.288-3a1 1 0 1 1 1.576-1.23l1.5 1.991 3.924-4.991a1 1 0 1 1 1.576 1.23l-4.712 6A.999.999 0 0 1 3.788 9z' class='spectrum-UIIcon--medium'/%3E%3C/svg%3E%0A");
@@ -176,10 +128,6 @@ merch-card.has-divider hr {
     margin-bottom: var(--consonant-merch-spacing-xs);
     height: 1px;
     border: none;
-}
-
-merch-card[variant="special-offers"] span[is="inline-price"][data-template="strikethrough"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
 }
 
 merch-card p, merch-card h3, merch-card h4 {
@@ -336,153 +284,10 @@ merch-card [slot="body-xl"] {
     color: var(--merch-color-grey-80);
 }
 
-merch-card[variant="plans"] [slot="description"] {
-    min-height: 84px;
-}
-
-merch-card[variant="catalog"] [slot="action-menu-content"] {
-    background-color: #000;
-    color: var(--color-white, #fff);
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    width: fit-content;
-    padding: var(--consonant-merch-spacing-xs);
-    border-radius: var(--consonant-merch-spacing-xxxs);
-    position: absolute;
-    top: 55px;
-    right: 15px;
-    line-height: var(--consonant-merch-card-body-line-height);
-}
-
-merch-card[variant="catalog"] [slot="action-menu-content"] ul {
-    padding-left: 0;
-    padding-bottom: var(--consonant-merch-spacing-xss);
-    margin-top: 0;
-    margin-bottom: 0;
-    list-style-position: inside;
-    list-style-type: '• ';
-}
-
-merch-card[variant="catalog"] [slot="action-menu-content"] ul li {
-    padding-left: 0;
-    line-height: var(--consonant-merch-card-body-line-height);
-}
-
-merch-card[variant="catalog"] [slot="action-menu-content"] ::marker {
-    margin-right: 0;
-}
-
-merch-card[variant="catalog"] [slot="action-menu-content"] p {
-    color: var(--color-white, #fff);
-}
-
-merch-card[variant="catalog"] [slot="action-menu-content"] a {
-    color: var(--consonant-merch-card-background-color);
-    text-decoration: underline;
-}
-
-merch-card[variant="catalog"] .payment-details {
-  font-size: var(--consonant-merch-card-body-font-size);
-  font-style: italic;
-  font-weight: 400;
-  line-height: var(--consonant-merch-card-body-line-height);
-}
-
-merch-card[variant="ccd-action"] .price-strikethrough {
-    font-size: 18px;
-}
-
-merch-card[variant="plans"] [slot="quantity-select"] {
-    display: flex;
-    justify-content: flex-start;
-    box-sizing: border-box;
-    width: 100%;
-    padding: var(--consonant-merch-spacing-xs);
-}
-
-merch-card[variant="twp"] div[class$='twp-badge'] {
-    padding: 4px 10px 5px 10px;
-}
-
-merch-card[variant="twp"] [slot="body-xs-top"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-    color: var(--merch-color-grey-80);
-}
-
-merch-card[variant="twp"] [slot="body-xs"] ul {
-    padding: 0;
-    margin: 0;
-}
-
-merch-card[variant="twp"] [slot="body-xs"] ul li {
-    list-style-type: none;
-    padding-left: 0;
-}
-
-merch-card[variant="twp"] [slot="body-xs"] ul li::before {
-    content: '·';
-    font-size: 20px;
-    padding-right: 5px;
-    font-weight: bold;
-}
-
-merch-card[variant="twp"] [slot="footer"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-    padding: var(--consonant-merch-spacing-s)
-    var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-xs);
-    color: var(--merch-color-grey-80);
-    display: flex;
-    flex-flow: wrap;
-}
-
-merch-card[variant='twp'] merch-quantity-select,
-merch-card[variant='twp'] merch-offer-select {
-    display: none;
-}
-
 [slot="cci-footer"] p,
 [slot="cct-footer"] p,
 [slot="cce-footer"] p {
     margin: 0;
-}
-
-/* mini compare chart card styles */
-
-merch-card[variant="mini-compare-chart"] [slot="heading-m"] {
-    padding: 0 var(--consonant-merch-spacing-s) 0;
-}
-
-merch-card[variant="mini-compare-chart"] [slot="body-m"] {
-    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s);
-}
-
-merch-card[variant="mini-compare-chart"] [is="inline-price"] {
-    display: inline-block;
-    min-height: 30px;
-    min-width: 1px;
-}
-
-merch-card[variant="mini-compare-chart"] [slot='callout-content'] {
-    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s) 0px;
-}
-
-merch-card[variant="mini-compare-chart"] [slot='callout-content'] [is="inline-price"] {
-    min-height: unset;
-}
-
-merch-card[variant="mini-compare-chart"] [slot="price-commitment"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    padding: 0 var(--consonant-merch-spacing-s);
-}
-
-merch-card[variant="mini-compare-chart"] [slot="price-commitment"] a {
-    display: inline-block;
-    height: 27px;
-}
-
-merch-card[variant="mini-compare-chart"] [slot="offers"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
 }
 
 merch-card [slot="promo-text"] {
@@ -493,111 +298,6 @@ merch-card [slot="promo-text"] {
     margin: 0;
     min-height: var(--consonant-merch-card-promo-text-height);
     padding: 0;
-}
-
-
-merch-card[variant="mini-compare-chart"] [slot="body-xxs"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s) 0;    
-}
-
-merch-card[variant="mini-compare-chart"] [slot="promo-text"] {
-    font-size: var(--consonant-merch-card-body-m-font-size);
-    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s) 0;
-}
-
-merch-card[variant="mini-compare-chart"] [slot="promo-text"] a {
-    text-decoration: underline;
-}
-
-merch-card[variant="mini-compare-chart"] .footer-row-icon {
-    display: flex;
-    place-items: center;
-}
-
-merch-card[variant="mini-compare-chart"] .footer-row-icon img {
-    max-width: initial;
-    width: var(--consonant-merch-card-mini-compare-chart-icon-size);
-    height: var(--consonant-merch-card-mini-compare-chart-icon-size);
-}
-
-merch-card[variant="mini-compare-chart"] .footer-row-cell {
-    border-top: 1px solid var(--consonant-merch-card-border-color);
-    display: flex;
-    gap: var(--consonant-merch-spacing-xs);
-    justify-content: start;
-    place-items: center;
-    padding: var(--consonant-merch-spacing-xs) var(--consonant-merch-spacing-s);
-}
-
-merch-card[variant="mini-compare-chart"] .footer-row-cell-description {
-    font-size: var(--consonant-merch-card-body-s-font-size);
-    line-height: var(--consonant-merch-card-body-s-line-height);
-}
-
-merch-card[variant="mini-compare-chart"] .footer-row-cell-description p {
-    color: var(--merch-color-grey-80);
-    vertical-align: bottom;
-}
-
-merch-card[variant="mini-compare-chart"] .footer-row-cell-description a {
-    color: var(--color-accent);
-    text-decoration: solid;
-}
-
-/* mini compare mobile */ 
-@media screen and ${MOBILE_LANDSCAPE} {
-  merch-card[variant="mini-compare-chart"] [slot='heading-m'] {
-    font-size: var(--consonant-merch-card-body-s-font-size);
-    line-height: var(--consonant-merch-card-body-s-line-height);
-  }
-
-  merch-card[variant="mini-compare-chart"] [slot='heading-m-price'] {
-    font-size: var(--consonant-merch-card-body-s-font-size);
-    line-height: var(--consonant-merch-card-body-s-line-height);
-  }
-
-  merch-card[variant="mini-compare-chart"] [slot='body-m'] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-  }
-  
-  merch-card[variant="mini-compare-chart"] [slot="promo-text"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-  }
-  merch-card[variant="mini-compare-chart"] .footer-row-cell-description {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-  }
-}
-
-/* mini compare tablet */
-@media screen and ${TABLET_DOWN} {
-  merch-card[variant="mini-compare-chart"] [slot='heading-m'] {
-    font-size: var(--consonant-merch-card-body-s-font-size);
-    line-height: var(--consonant-merch-card-body-s-line-height);
-  }
-
-  merch-card[variant="mini-compare-chart"] [slot='heading-m-price'] {
-    font-size: var(--consonant-merch-card-body-s-font-size);
-    line-height: var(--consonant-merch-card-body-s-line-height);
-  }
-
-  merch-card[variant="mini-compare-chart"] [slot='body-m'] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-  }
-
-  merch-card[variant="mini-compare-chart"] [slot="promo-text"] {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-  }
-  
-  merch-card[variant="mini-compare-chart"] .footer-row-cell-description {
-    font-size: var(--consonant-merch-card-body-xs-font-size);
-    line-height: var(--consonant-merch-card-body-xs-line-height);
-  }
 }
 
 div[slot="footer"] {
@@ -622,451 +322,6 @@ div[slot='bg-image'] img {
     object-fit: cover;
     border-top-left-radius: 16px;
     border-top-right-radius: 16px;
-}
-
-/* Mobile */
-@media screen and ${MOBILE_LANDSCAPE} {
-    :root {
-        --consonant-merch-card-mini-compare-chart-width: 302px;
-        --consonant-merch-card-segment-width: 276px;
-        --consonant-merch-card-mini-compare-chart-wide-width: 302px;
-        --consonant-merch-card-special-offers-width: 302px;
-        --consonant-merch-card-twp-width: 300px;
-    }
-}
-
-
-/* Tablet */
-@media screen and ${TABLET_UP} {
-    :root {
-        --consonant-merch-card-catalog-width: 302px;
-        --consonant-merch-card-plans-width: 302px;
-        --consonant-merch-card-segment-width: 276px;
-        --consonant-merch-card-mini-compare-chart-width: 302px;
-        --consonant-merch-card-mini-compare-chart-wide-width: 302px;
-        --consonant-merch-card-special-offers-width: 302px;
-        --consonant-merch-card-twp-width: 268px;
-    }
-}
-
-/* desktop */
-@media screen and ${DESKTOP_UP} {
-    :root {
-        --consonant-merch-card-catalog-width: 276px;
-        --consonant-merch-card-plans-width: 276px;
-        --consonant-merch-card-segment-width: 302px;
-        --consonant-merch-card-inline-heading-width: 378px;
-        --consonant-merch-card-product-width: 378px;
-        --consonant-merch-card-image-width: 378px;
-        --consonant-merch-card-mini-compare-chart-width: 378px;
-        --consonant-merch-card-mini-compare-chart-wide-width: 484px;
-        --consonant-merch-card-twp-width: 268px;
-    }
-}
-
-/* supported cards */
-/* grid style for plans */
-.one-merch-card.plans,
-.two-merch-cards.plans,
-.three-merch-cards.plans,
-.four-merch-cards.plans {
-    grid-template-columns: var(--consonant-merch-card-plans-width);
-}
-
-/* Tablet */
-@media screen and ${TABLET_UP} {
-    .two-merch-cards.plans,
-    .three-merch-cards.plans,
-    .four-merch-cards.plans {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-plans-width));
-    }
-}
-
-/* desktop */
-@media screen and ${DESKTOP_UP} {
-    .three-merch-cards.plans,
-    .four-merch-cards.plans {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-plans-width));
-    }
-}
-
-/* Large desktop */
-    @media screen and ${LARGE_DESKTOP} {
-    .four-merch-cards.plans {
-        grid-template-columns: repeat(4, var(--consonant-merch-card-plans-width));
-    }
-}
-
-
-/* grid style for catalog */
-.one-merch-card.catalog,
-.two-merch-cards.catalog,
-.three-merch-cards.catalog,
-.four-merch-cards.catalog {
-    grid-template-columns: var(--consonant-merch-card-catalog-width);
-}
-
-/* Tablet */
-@media screen and ${TABLET_UP} {
-    .two-merch-cards.catalog,
-    .three-merch-cards.catalog,
-    .four-merch-cards.catalog {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-catalog-width));
-    }
-}
-
-/* desktop */
-@media screen and ${DESKTOP_UP} {
-    .three-merch-cards.catalog,
-    .four-merch-cards.catalog {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-catalog-width));
-    }
-}
-
-/* Large desktop */
-    @media screen and ${LARGE_DESKTOP} {
-    .four-merch-cards.catalog {
-        grid-template-columns: repeat(4, var(--consonant-merch-card-catalog-width));
-    }
-}
-
-
-/* grid style for special-offers */
-.one-merch-card.special-offers,
-.two-merch-cards.special-offers,
-.three-merch-cards.special-offers,
-.four-merch-cards.special-offers {
-    grid-template-columns: minmax(300px, var(--consonant-merch-card-special-offers-width));
-}
-
-/* Tablet */
-@media screen and ${TABLET_UP} {
-    .two-merch-cards.special-offers,
-    .three-merch-cards.special-offers,
-    .four-merch-cards.special-offers {
-        grid-template-columns: repeat(2, minmax(300px, var(--consonant-merch-card-special-offers-width)));
-    }
-}
-
-/* desktop */
-@media screen and ${DESKTOP_UP} {
-    .three-merch-cards.special-offers,
-    .four-merch-cards.special-offers {
-        grid-template-columns: repeat(3, minmax(300px, var(--consonant-merch-card-special-offers-width)));
-    }
-}
-
-@media screen and ${LARGE_DESKTOP} {
-    .four-merch-cards.special-offers {
-        grid-template-columns: repeat(4, minmax(300px, var(--consonant-merch-card-special-offers-width)));
-    }
-}
-
-
-/* grid style for image */
-.one-merch-card.image,
-.two-merch-cards.image,
-.three-merch-cards.image,
-.four-merch-cards.image {
-    grid-template-columns: var(--consonant-merch-card-image-width);
-}
-
-/* Tablet */
-@media screen and ${TABLET_UP} {
-    .two-merch-cards.image,
-    .three-merch-cards.image,
-    .four-merch-cards.image {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-image-width));
-    }
-}
-
-/* desktop */
-@media screen and ${DESKTOP_UP} {
-    .three-merch-cards.image,
-    .four-merch-cards.image {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-image-width));
-    }
-}
-
-/* Large desktop */
-    @media screen and ${LARGE_DESKTOP} {
-    .four-merch-cards.image {
-        grid-template-columns: repeat(4, var(--consonant-merch-card-image-width));
-    }
-}
-
-
-/* grid style for segment */
-.one-merch-card.segment,
-.two-merch-cards.segment,
-.three-merch-cards.segment,
-.four-merch-cards.segment {
-    grid-template-columns: minmax(276px, var(--consonant-merch-card-segment-width));
-}
-
-/* Tablet */
-@media screen and ${TABLET_UP} {
-    .two-merch-cards.segment,
-    .three-merch-cards.segment,
-    .four-merch-cards.segment {
-        grid-template-columns: repeat(2, minmax(276px, var(--consonant-merch-card-segment-width)));
-    }
-}
-
-/* desktop */
-@media screen and ${DESKTOP_UP} {
-    .three-merch-cards.segment {
-        grid-template-columns: repeat(3, minmax(276px, var(--consonant-merch-card-segment-width)));
-    }
-
-    .four-merch-cards.segment {
-        grid-template-columns: repeat(4, minmax(276px, var(--consonant-merch-card-segment-width)));
-    }
-}
-
-
-/* grid style for product */
-.one-merch-card.product,
-.two-merch-cards.product,
-.three-merch-cards.product,
-.four-merch-cards.product {
-    grid-template-columns: var(--consonant-merch-card-product-width);
-}
-
-/* Tablet */
-@media screen and ${TABLET_UP} {
-    .two-merch-cards.product,
-    .three-merch-cards.product,
-    .four-merch-cards.product {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-product-width));
-    }
-}
-
-/* desktop */
-@media screen and ${DESKTOP_UP} {
-    .three-merch-cards.product,
-    .four-merch-cards.product {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-product-width));
-    }
-}
-
-/* Large desktop */
-    @media screen and ${LARGE_DESKTOP} {
-    .four-merch-cards.product {
-        grid-template-columns: repeat(4, var(--consonant-merch-card-product-width));
-    }
-}
-
-/* grid style for twp */
-.one-merch-card.twp,
-.two-merch-cards.twp,
-.three-merch-cards.twp {
-    grid-template-columns: var(--consonant-merch-card-image-width);
-}
-
-/* Tablet */
-@media screen and ${TABLET_UP} {
-    .one-merch-card.twp,
-    .two-merch-cards.twp {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-twp-width));
-    }
-    .three-merch-cards.twp {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-twp-width));
-    }
-}
-
-/* desktop */
-@media screen and ${DESKTOP_UP} {
-    .one-merch-card.twp
-    .two-merch-cards.twp {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-twp-width));
-    }
-    .three-merch-cards.twp {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-twp-width));
-    }
-}
-
-/* Large desktop */
-    @media screen and ${LARGE_DESKTOP} {
-        .one-merch-card.twp
-        .two-merch-cards.twp {
-            grid-template-columns: repeat(2, var(--consonant-merch-card-twp-width));
-        }
-        .three-merch-cards.twp {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-twp-width));
-    }
-}
-
-/* Mobile */
-@media screen and ${MOBILE_LANDSCAPE} {
-    .one-merch-card.twp,
-    .two-merch-cards.twp,
-    .three-merch-cards.twp {
-        grid-template-columns: repeat(1, var(--consonant-merch-card-twp-mobile-width));
-    }
-}
-
-/* grid style for inline-heading */
-.one-merch-card.inline-heading,
-.two-merch-cards.inline-heading,
-.three-merch-cards.inline-heading,
-.four-merch-cards.inline-heading {
-    grid-template-columns: var(--consonant-merch-card-inline-heading-width);
-}
-
-/* Tablet */
-@media screen and ${TABLET_UP} {
-    .two-merch-cards.inline-heading,
-    .three-merch-cards.inline-heading,
-    .four-merch-cards.inline-heading {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-inline-heading-width));
-    }
-}
-
-/* desktop */
-@media screen and ${DESKTOP_UP} {
-    .three-merch-cards.inline-heading,
-    .four-merch-cards.inline-heading {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-inline-heading-width));
-    }
-}
-
-/* Large desktop */
-    @media screen and ${LARGE_DESKTOP} {
-    .four-merch-cards.inline-heading {
-        grid-template-columns: repeat(4, var(--consonant-merch-card-inline-heading-width));
-    }
-}
-
-/* grid style for ccd-action */
-.one-merch-card.ccd-action,
-.two-merch-cards.ccd-action,
-.three-merch-cards.ccd-action,
-.four-merch-cards.ccd-action {
-    grid-template-columns: var(--consonant-merch-card-ccd-action-width);
-}
-
-/* Tablet */
-@media screen and ${TABLET_UP} {
-    .two-merch-cards.ccd-action,
-    .three-merch-cards.ccd-action,
-    .four-merch-cards.ccd-action {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-ccd-action-width));
-    }
-}
-
-/* desktop */
-@media screen and ${DESKTOP_UP} {
-    .three-merch-cards.ccd-action,
-    .four-merch-cards.ccd-action {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-ccd-action-width));
-    }
-}
-
-/* Large desktop */
-    @media screen and ${LARGE_DESKTOP} {
-    .four-merch-cards.ccd-action {
-        grid-template-columns: repeat(4, var(--consonant-merch-card-ccd-action-width));
-    }
-}
-
-/* grid style for mini-compare-chart */
-.one-merch-card.mini-compare-chart {
-    grid-template-columns: var(--consonant-merch-card-mini-compare-chart-wide-width);
-    gap: var(--consonant-merch-spacing-xs);
-}
-
-.two-merch-cards.mini-compare-chart,
-.three-merch-cards.mini-compare-chart,
-.four-merch-cards.mini-compare-chart {
-    grid-template-columns: repeat(2, var(--consonant-merch-card-mini-compare-chart-width));
-    gap: var(--consonant-merch-spacing-xs);
-}
-
-@media screen and ${MOBILE_LANDSCAPE} {
-    .two-merch-cards.mini-compare-chart,
-    .three-merch-cards.mini-compare-chart,
-    .four-merch-cards.mini-compare-chart {
-        grid-template-columns: var(--consonant-merch-card-mini-compare-chart-width);
-        gap: var(--consonant-merch-spacing-xs);
-    }
-}
-
-@media screen and ${TABLET_DOWN} {
-    .three-merch-cards.mini-compare-chart merch-card [slot="footer"] a,
-    .four-merch-cards.mini-compare-chart merch-card [slot="footer"] a {
-        flex: 1;
-    }
-}
-
-/* Tablet */
-@media screen and ${TABLET_UP} {
-    .two-merch-cards.mini-compare-chart {
-        grid-template-columns: repeat(2, minmax(var(--consonant-merch-card-mini-compare-chart-width), var(--consonant-merch-card-mini-compare-chart-wide-width)));
-        gap: var(--consonant-merch-spacing-m);
-    }
-
-    .three-merch-cards.mini-compare-chart,
-    .four-merch-cards.mini-compare-chart {
-        grid-template-columns: repeat(2, minmax(var(--consonant-merch-card-mini-compare-chart-width), var(--consonant-merch-card-mini-compare-chart-wide-width)));
-    }
-}
-
-/* desktop */
-@media screen and ${DESKTOP_UP} {
-    .one-merch-card.mini-compare-chart {
-        grid-template-columns: var(--consonant-merch-card-mini-compare-chart-wide-width);
-    }
-
-    .two-merch-cards.mini-compare-chart {
-        grid-template-columns: repeat(2, var(--consonant-merch-card-mini-compare-chart-wide-width));
-        gap: var(--consonant-merch-spacing-m);
-    }
-
-    .three-merch-cards.mini-compare-chart,
-    .four-merch-cards.mini-compare-chart {
-        grid-template-columns: repeat(3, var(--consonant-merch-card-mini-compare-chart-width));
-        gap: var(--consonant-merch-spacing-m);
-    }
-}
-
-@media screen and ${LARGE_DESKTOP} {
-    .four-merch-cards.mini-compare-chart {
-        grid-template-columns: repeat(4, var(--consonant-merch-card-mini-compare-chart-width));
-    }
-}
-
-/* mini-compare card footer rows */
-merch-card .footer-row-cell:nth-child(1) {
-    min-height: var(--consonant-merch-card-footer-row-1-min-height);
-}
-
-merch-card .footer-row-cell:nth-child(2) {
-    min-height: var(--consonant-merch-card-footer-row-2-min-height);
-}
-
-merch-card .footer-row-cell:nth-child(3) {
-    min-height: var(--consonant-merch-card-footer-row-3-min-height);
-}
-
-merch-card .footer-row-cell:nth-child(4) {
-    min-height: var(--consonant-merch-card-footer-row-4-min-height);
-}
-
-merch-card .footer-row-cell:nth-child(5) {
-    min-height: var(--consonant-merch-card-footer-row-5-min-height);
-}
-
-merch-card .footer-row-cell:nth-child(6) {
-    min-height: var(--consonant-merch-card-footer-row-6-min-height);
-}
-
-merch-card .footer-row-cell:nth-child(7) {
-    min-height: var(--consonant-merch-card-footer-row-7-min-height);
-}
-
-merch-card .footer-row-cell:nth-child(8) {
-    min-height: var(--consonant-merch-card-footer-row-8-min-height);
 }
 
 span[is="inline-price"][data-template='strikethrough'] {
