+++ libs/features/mas/web-components/src/merch-card.js
@@ -1,6 +1,6 @@
 import { html, LitElement, nothing } from 'lit';
 import { sizeStyles, styles } from './merch-card.css.js';
-import { isMobile, isMobileOrTablet } from './utils.js';
+import { getVariantLayout, getVariantStyles } from './variants/variants.js';
 
 import './global.css.js';
 import {
@@ -8,19 +8,11 @@ import {
     EVENT_MERCH_OFFER_SELECT_READY,
     EVENT_MERCH_QUANTITY_SELECTOR_CHANGE,
     EVENT_MERCH_STORAGE_CHANGE,
-    EVENT_MERCH_CARD_ACTION_MENU_TOGGLE,
 } from './constants.js';
 import { getTextNodes } from './utils.js';
 
 export const MERCH_CARD_NODE_NAME = 'MERCH-CARD';
 export const MERCH_CARD = 'merch-card';
-
-const FOOTER_ROW_MIN_HEIGHT = 32; // as per the XD.
-
-const MINI_COMPARE_CHART = 'mini-compare-chart';
-
-const getRowMinHeightPropertyName = (index) =>
-    `--consonant-merch-card-footer-row-${index}-min-height`;
 
 export class MerchCard extends LitElement {
     static properties = {
@@ -90,10 +82,11 @@ export class MerchCard extends LitElement {
         merchOffer: { type: Object },
     };
 
-    static styles = [styles, ...sizeStyles()];
+    static styles = [styles, getVariantStyles(), ...sizeStyles()];
 
     customerSegment;
     marketSegment;
+    variantLayout; 
 
     constructor() {
         super();
@@ -101,8 +94,6 @@ export class MerchCard extends LitElement {
         this.types = '';
         this.selected = false;
     }
-
-    #container;
 
     updated(changedProperties) {
         if (
@@ -120,14 +111,13 @@ export class MerchCard extends LitElement {
                 (price) => !price.closest('[slot="callout-content"]'),
             );
             await Promise.all(prices.map((price) => price.onceSettled()));
-            this.adjustTitleWidth();
-            if (!isMobile()) {
-              this.adjustMiniCompareBodySlots();
-              this.adjustMiniCompareFooterRows();
-            } else {
-              this.removeEmptyRows();
-            }
+            this.variantLayout.postCardUpdateHook(this);          
         });
+    }
+
+    render() {
+      if (!this.isConnected || this.style.display === 'none') return;
+      return this.variantLayout.renderLayout();
     }
 
     get computedBorderStyle() {
@@ -139,74 +129,8 @@ export class MerchCard extends LitElement {
         return '';
     }
 
-    get evergreen() {
-        return this.classList.contains('intro-pricing');
-    }
-
-    get stockCheckbox() {
-        return this.checkboxLabel
-            ? html`<label id="stock-checkbox">
-                    <input type="checkbox" @change=${this.toggleStockOffer}></input>
-                    <span></span>
-                    ${this.checkboxLabel}
-                </label>`
-            : '';
-    }
-
-    get cardImage() {
-        return html` <div class="image">
-            <slot name="bg-image"></slot>
-            ${this.badge}
-        </div>`;
-    }
-
-    get secureLabelFooter() {
-        const secureLabel = this.secureLabel
-            ? html`<span class="secure-transaction-label"
-                  >${this.secureLabel}</span
-              >`
-            : '';
-        return html`<footer>${secureLabel}<slot name="footer"></slot></footer>`;
-    }
-
-    get miniCompareFooter() {
-        const secureLabel = this.secureLabel
-            ? html`<slot name="secure-transaction-label">
-                  <span class="secure-transaction-label"
-                      >${this.secureLabel}</span
-                  ></slot
-              >`
-            : html`<slot name="secure-transaction-label"></slot>`;
-        return html`<footer>${secureLabel}<slot name="footer"></slot></footer>`;
-    }
-
-    get badge() {
-        let additionalStyles;
-        if (!this.badgeBackgroundColor || !this.badgeColor || !this.badgeText) {
-            return;
-        }
-        if (this.evergreen) {
-            additionalStyles = `border: 1px solid ${this.badgeBackgroundColor}; border-right: none;`;
-        }
-        return html`
-            <div
-                id="badge"
-                class="${this.variant}-badge"
-                style="background-color: ${this.badgeBackgroundColor};
-                    color: ${this.badgeColor};
-                    ${additionalStyles}"
-            >
-                ${this.badgeText}
-            </div>
-        `;
-    }
-
     get badgeElement() {
         return this.shadowRoot.getElementById('badge');
-    }
-
-    getContainer() {
-        return this.closest('[class*="-merch-cards"]') ?? this.parentElement;
     }
 
     get headingmMSlot() {
@@ -253,27 +177,6 @@ export class MerchCard extends LitElement {
         }
     }
 
-    toggleActionMenu(e) {
-        const retract = e?.type === 'mouseleave' ? true : undefined;
-        const actionMenuContentSlot = this.shadowRoot.querySelector(
-            'slot[name="action-menu-content"]',
-        );
-        if (!actionMenuContentSlot) return;
-        if (!retract) {
-            this.dispatchEvent(
-                new CustomEvent(EVENT_MERCH_CARD_ACTION_MENU_TOGGLE, {
-                    bubbles: true,
-                    composed: true,
-                    detail: {
-                        card: this.name,
-                        type: 'action-menu',
-                    },
-                }),
-            );
-        }
-        actionMenuContentSlot.classList.toggle('hidden', retract);
-    }
-
     handleQuantitySelection(event) {
         const elements = this.checkoutLinks;
         for (const element of elements) {
@@ -281,12 +184,8 @@ export class MerchCard extends LitElement {
         }
     }
 
-    get titleElement() {
-        const heading =
-            this.variant === 'special-offers'
-                ? this.querySelector('[slot="detail-m"]')
-                : this.querySelector('[slot="heading-xs"]');
-        return heading;
+    get titleElement() {        
+        return this.querySelector(this.variantLayout.titleSelector);
     }
 
     get title() {
@@ -322,206 +221,8 @@ export class MerchCard extends LitElement {
         return this.textContent.match(new RegExp(text, 'i')) !== null;
     }
 
-    render() {
-        if (!this.isConnected || this.style.display === 'none') return;
-        switch (this.variant) {
-            case 'special-offers':
-                return this.renderSpecialOffer();
-            case 'segment':
-                return this.renderSegment();
-            case 'plans':
-                return this.renderPlans();
-            case 'catalog':
-                return this.renderCatalog();
-            case 'image':
-                return this.renderImage();
-            case 'product':
-                return this.renderProduct();
-            case 'inline-heading':
-                return this.renderInlineHeading();
-            case MINI_COMPARE_CHART:
-                return this.renderMiniCompareChart();
-            case 'ccd-action':
-                return this.renderCcdAction();
-            case 'twp':
-                return this.renderTwp();
-            default:
-                // this line should never hit, to check.
-                return this.renderProduct();
-        }
-    }
-
-    renderSpecialOffer() {
-        return html`${this.cardImage}
-            <div class="body">
-                <slot name="detail-m"></slot>
-                <slot name="heading-xs"></slot>
-                <slot name="body-xs"></slot>
-            </div>
-            ${this.evergreen
-                ? html`
-                      <div
-                          class="detail-bg-container"
-                          style="background: ${this['detailBg']}"
-                      >
-                          <slot name="detail-bg"></slot>
-                      </div>
-                  `
-                : html`
-                      <hr />
-                      ${this.secureLabelFooter}
-                  `}`;
-    }
-
-    get promoBottom() {
-      return this.classList.contains('promo-bottom');
-    }
-
     get startingAt() {
       return this.classList.contains('starting-at');
-    }
-
-    renderSegment() {
-        return html` ${this.badge}
-            <div class="body">
-                <slot name="heading-xs"></slot>
-                <slot name="body-xxs"></slot>
-                ${!this.promoBottom ? html`<slot name="promo-text"></slot><slot name="callout-content"></slot>` : ''}
-                <slot name="body-xs"></slot>
-                ${this.promoBottom ? html`<slot name="promo-text"></slot><slot name="callout-content"></slot>` : ''}
-            </div>
-            <hr />
-            ${this.secureLabelFooter}`;
-    }
-
-    renderPlans() {
-        return html` ${this.badge}
-            <div class="body">
-                <slot name="icons"></slot>
-                <slot name="heading-xs"></slot>
-                <slot name="heading-m"></slot>
-                <slot name="body-xxs"></slot>
-                ${!this.promoBottom ? html`<slot name="promo-text"></slot><slot name="callout-content"></slot> ` : ''}
-                <slot name="body-xs"></slot>
-                ${this.promoBottom ? html`<slot name="promo-text"></slot><slot name="callout-content"></slot> ` : ''}  
-                ${this.stockCheckbox}
-            </div>
-            <slot name="quantity-select"></slot>
-            ${this.secureLabelFooter}`;
-    }
-
-    renderCatalog() {
-        return html` <div class="body">
-                <div class="top-section">
-                    <slot name="icons"></slot> ${this.badge}
-                    <div
-                        class="action-menu
-                        ${isMobileOrTablet() && this.actionMenu
-                            ? 'always-visible'
-                            : ''}
-                        ${!this.actionMenu ? 'hidden' : 'invisible'}"
-                        @click="${this.toggleActionMenu}"
-                    ></div>
-                </div>
-                <slot
-                    name="action-menu-content"
-                    class="action-menu-content
-                    ${!this.actionMenuContent ? 'hidden' : ''}"
-                    >${this.actionMenuContent}</slot
-                >
-                <slot name="heading-xs"></slot>
-                <slot name="heading-m"></slot>
-                <slot name="body-xxs"></slot>
-                ${!this.promoBottom
-                    ? html`<slot name="promo-text"></slot
-                          ><slot name="callout-content"></slot>`
-                    : ''}
-                <slot name="body-xs"></slot>
-                ${this.promoBottom
-                    ? html`<slot name="promo-text"></slot
-                          ><slot name="callout-content"></slot>`
-                    : ''}
-            </div>
-            ${this.secureLabelFooter}`;
-    }
-
-    renderImage() {
-        return html`${this.cardImage}
-            <div class="body">
-                <slot name="icons"></slot>
-                <slot name="heading-xs"></slot>
-                <slot name="body-xxs"></slot>
-                ${this.promoBottom ? html`<slot name="body-xs"></slot><slot name="promo-text"></slot>` : html`<slot name="promo-text"></slot><slot name="body-xs"></slot>`}
-            </div>
-            ${this.evergreen
-                ? html`
-                      <div
-                          class="detail-bg-container"
-                          style="background: ${this['detailBg']}"
-                      >
-                          <slot name="detail-bg"></slot>
-                      </div>
-                  `
-                : html`
-                      <hr />
-                      ${this.secureLabelFooter}
-                  `}`;
-    }
-
-    renderInlineHeading() {
-        return html` ${this.badge}
-            <div class="body">
-                <div class="top-section">
-                    <slot name="icons"></slot>
-                    <slot name="heading-xs"></slot>
-                </div>
-                <slot name="body-xs"></slot>
-            </div>
-            ${!this.customHr ? html`<hr />` : ''} ${this.secureLabelFooter}`;
-    }
-
-    renderProduct() {
-        return html` ${this.badge}
-            <div class="body">
-                <slot name="icons"></slot>
-                <slot name="heading-xs"></slot>
-                <slot name="body-xxs"></slot>
-                ${!this.promoBottom ? html`<slot name="promo-text"></slot><slot name="callout-content"></slot>` : ''}
-                <slot name="body-xs"></slot>
-                ${this.promoBottom ? html`<slot name="promo-text"></slot><slot name="callout-content"></slot>` : ''}
-            </div>
-            ${this.secureLabelFooter}`;
-    }
-
-    renderMiniCompareChart() {
-        // Define the HTML structure for the 'mini-compare-chart' variant here
-        const { badge } = this;
-        return html` <div class="top-section${badge ? ' badge' : ''}">
-                <slot name="icons"></slot> ${badge}
-            </div>
-            <slot name="heading-m"></slot>
-            <slot name="body-m"></slot>
-            <slot name="heading-m-price"></slot>
-            <slot name="body-xxs"></slot>
-            <slot name="price-commitment"></slot>
-            <slot name="offers"></slot>
-            <slot name="promo-text"></slot>
-            <slot name="callout-content"></slot>
-            ${this.miniCompareFooter}
-            <slot name="footer-rows"><slot name="body-s"></slot></slot>`;
-    }
-
-    renderTwp() {
-        return html`${this.badge}
-            <div class="top-section">
-                <slot name="icons"></slot>
-                <slot name="heading-xs"></slot>
-                <slot name="body-xs-top"></slot>
-            </div>
-            <div class="body">
-                <slot name="body-xs"></slot>
-            </div>
-            <footer><slot name="footer"></slot></footer>`;
     }
 
     get defaultSlot() {
@@ -532,22 +233,11 @@ export class MerchCard extends LitElement {
         return html`<slot></slot>`;
     }
 
-    renderCcdAction() {
-        return html` <div class="body">
-            <slot name="icons"></slot> ${this.badge}
-            <slot name="heading-xs"></slot>
-            <slot name="heading-m"></slot>
-            ${this.promoBottom ? html`<slot name="body-xs"></slot><slot name="promo-text"></slot>` : html`<slot name="promo-text"></slot><slot name="body-xs"></slot>`}
-            <footer><slot name="footer"></slot></footer>
-            ${this.defaultSlot}
-        </div>`;
-    }
-
     connectedCallback() {
         super.connectedCallback();
-        this.#container = this.getContainer();
+        this.variantLayout = getVariantLayout(this);
+        this.variantLayout.connectedCallbackHook();
         this.setAttribute('tabindex', this.getAttribute('tabindex') ?? '0');
-        this.addEventListener('mouseleave', this.toggleActionMenu);
         this.addEventListener(
             EVENT_MERCH_QUANTITY_SELECTOR_CHANGE,
             this.handleQuantitySelection,
@@ -569,6 +259,7 @@ export class MerchCard extends LitElement {
 
     disconnectedCallback() {
         super.disconnectedCallback();
+        this.variantLayout.disconnectedCallbackHook();
 
         this.removeEventListener(
             EVENT_MERCH_QUANTITY_SELECTOR_CHANGE,
@@ -597,116 +288,6 @@ export class MerchCard extends LitElement {
     }
 
     // custom methods
-
-    updateMiniCompareElementMinHeight(el, name) {
-        const elMinHeightPropertyName = `--consonant-merch-card-mini-compare-${name}-height`;
-        const height = Math.max(
-            0,
-            parseInt(window.getComputedStyle(el).height) || 0,
-        );
-        const maxMinHeight =
-            parseInt(
-                this.#container.style.getPropertyValue(elMinHeightPropertyName),
-            ) || 0;
-        if (height > maxMinHeight) {
-            this.#container.style.setProperty(
-                elMinHeightPropertyName,
-                `${height}px`,
-            );
-        }
-    }
-
-    async adjustTitleWidth() {
-        if (!['segment', 'plans'].includes(this.variant)) return;
-        const cardWidth = this.getBoundingClientRect().width;
-        const badgeWidth =
-            this.badgeElement?.getBoundingClientRect().width || 0;
-        if (cardWidth === 0 || badgeWidth === 0) return;
-        this.style.setProperty(
-            '--consonant-merch-card-heading-xs-max-width',
-            `${Math.round(cardWidth - badgeWidth - 16)}px`, // consonant-merch-spacing-xs
-        );
-    }
-
-    async adjustMiniCompareBodySlots() {
-        if (this.variant !== MINI_COMPARE_CHART) return;
-        if (this.getBoundingClientRect().width === 0) return;
-
-        this.updateMiniCompareElementMinHeight(
-            this.shadowRoot.querySelector('.top-section'),
-            'top-section',
-        );
-
-        const slots = [
-            'heading-m',
-            'body-m',
-            'heading-m-price',
-            'price-commitment',
-            'offers',
-            'promo-text',
-            'callout-content',
-            'secure-transaction-label',
-        ];
-
-        slots.forEach((slot) =>
-            this.updateMiniCompareElementMinHeight(
-                this.shadowRoot.querySelector(`slot[name="${slot}"]`),
-                slot,
-            ),
-        );
-        this.updateMiniCompareElementMinHeight(
-            this.shadowRoot.querySelector('footer'),
-            'footer',
-        );
-
-        const badge = this.shadowRoot.querySelector(
-            '.mini-compare-chart-badge',
-        );
-        if (badge && badge.textContent !== '') {
-            this.#container.style.setProperty(
-                '--consonant-merch-card-mini-compare-top-section-mobile-height',
-                '32px',
-            );
-        }
-    }
-
-    adjustMiniCompareFooterRows() {
-        if (this.variant !== MINI_COMPARE_CHART) return;
-        if (this.getBoundingClientRect().width === 0) return;
-        const footerRows = this.querySelector('[slot="footer-rows"]');
-        [...footerRows.children].forEach((el, index) => {
-            const height = Math.max(
-                FOOTER_ROW_MIN_HEIGHT,
-                parseInt(window.getComputedStyle(el).height) || 0,
-            );
-            const maxMinHeight =
-                parseInt(
-                    this.#container.style.getPropertyValue(
-                        getRowMinHeightPropertyName(index + 1),
-                    ),
-                ) || 0;
-            if (height > maxMinHeight) {
-                this.#container.style.setProperty(
-                    getRowMinHeightPropertyName(index + 1),
-                    `${height}px`,
-                );
-            }
-        });
-    }
-
-    removeEmptyRows() {
-      if (this.variant !== MINI_COMPARE_CHART) return;
-      const footerRows = this.querySelectorAll('.footer-row-cell');
-      footerRows.forEach((row) => {
-          const rowDescription = row.querySelector('.footer-row-cell-description');
-          if (rowDescription) {
-              const isEmpty = !rowDescription.textContent.trim();
-              if (isEmpty) {
-                  row.remove();
-              }
-          }
-      });
-  }
 
     get storageOptions() {
         return this.querySelector('sp-radio-group#storage');
