+++ libs/features/mas/web-components/src/variants/catalog.css.js
@@ -1,0 +1,88 @@
+import { TABLET_UP, DESKTOP_UP, LARGE_DESKTOP,} from '../media.js';
+
+export const CSS = `
+:root {
+  --consonant-merch-card-catalog-width: 276px;
+  --consonant-merch-card-catalog-icon-size: 40px;
+}
+.one-merch-card.catalog,
+.two-merch-cards.catalog,
+.three-merch-cards.catalog,
+.four-merch-cards.catalog {
+    grid-template-columns: var(--consonant-merch-card-catalog-width);
+}
+
+@media screen and ${TABLET_UP} {
+    :root {
+      --consonant-merch-card-catalog-width: 302px;
+    }
+
+    .two-merch-cards.catalog,
+    .three-merch-cards.catalog,
+    .four-merch-cards.catalog {
+        grid-template-columns: repeat(2, var(--consonant-merch-card-catalog-width));
+    }
+}
+
+@media screen and ${DESKTOP_UP} {
+    :root {
+      --consonant-merch-card-catalog-width: 276px;
+    }
+    .three-merch-cards.catalog,
+    .four-merch-cards.catalog {
+        grid-template-columns: repeat(3, var(--consonant-merch-card-catalog-width));
+    }
+}
+
+@media screen and ${LARGE_DESKTOP} {
+    .four-merch-cards.catalog {
+        grid-template-columns: repeat(4, var(--consonant-merch-card-catalog-width));
+    }
+}
+
+merch-card[variant="catalog"] [slot="action-menu-content"] {
+  background-color: #000;
+  color: var(--color-white, #fff);
+  font-size: var(--consonant-merch-card-body-xs-font-size);
+  width: fit-content;
+  padding: var(--consonant-merch-spacing-xs);
+  border-radius: var(--consonant-merch-spacing-xxxs);
+  position: absolute;
+  top: 55px;
+  right: 15px;
+  line-height: var(--consonant-merch-card-body-line-height);
+}
+
+merch-card[variant="catalog"] [slot="action-menu-content"] ul {
+  padding-left: 0;
+  padding-bottom: var(--consonant-merch-spacing-xss);
+  margin-top: 0;
+  margin-bottom: 0;
+  list-style-position: inside;
+  list-style-type: '• ';
+}
+
+merch-card[variant="catalog"] [slot="action-menu-content"] ul li {
+  padding-left: 0;
+  line-height: var(--consonant-merch-card-body-line-height);
+}
+
+merch-card[variant="catalog"] [slot="action-menu-content"] ::marker {
+  margin-right: 0;
+}
+
+merch-card[variant="catalog"] [slot="action-menu-content"] p {
+  color: var(--color-white, #fff);
+}
+
+merch-card[variant="catalog"] [slot="action-menu-content"] a {
+  color: var(--consonant-merch-card-background-color);
+  text-decoration: underline;
+}
+
+merch-card[variant="catalog"] .payment-details {
+  font-size: var(--consonant-merch-card-body-font-size);
+  font-style: italic;
+  font-weight: 400;
+  line-height: var(--consonant-merch-card-body-line-height);
+}`;
