+++ libs/features/mas/web-components/src/variants/image.css.js
@@ -1,0 +1,38 @@
+import { TABLET_UP, DESKTOP_UP, LARGE_DESKTOP,} from '../media.js';
+export const CSS = `
+:root {
+  --consonant-merch-card-image-width: 300px;
+  --consonant-merch-card-image-height: 180px;
+}
+
+.one-merch-card.image,
+.two-merch-cards.image,
+.three-merch-cards.image,
+.four-merch-cards.image {
+  grid-template-columns: var(--consonant-merch-card-image-width);
+}
+
+@media screen and ${TABLET_UP} {
+  .two-merch-cards.image,
+  .three-merch-cards.image,
+  .four-merch-cards.image {
+      grid-template-columns: repeat(2, var(--consonant-merch-card-image-width));
+  }
+}
+
+@media screen and ${DESKTOP_UP} {
+  :root {
+    --consonant-merch-card-image-width: 378px;
+  }
+  .three-merch-cards.image,
+  .four-merch-cards.image {
+      grid-template-columns: repeat(3, var(--consonant-merch-card-image-width));
+  }
+}
+
+@media screen and ${LARGE_DESKTOP} {
+  .four-merch-cards.image {
+      grid-template-columns: repeat(4, var(--consonant-merch-card-image-width));
+  }
+}
+`;
