+++ libs/features/mas/web-components/src/variants/inline-heading.css.js
@@ -1,0 +1,35 @@
+import { TABLET_UP, DESKTOP_UP, LARGE_DESKTOP,} from '../media.js';
+
+export const CSS = `
+--consonant-merch-card-inline-heading-width: 300px;
+
+.one-merch-card.inline-heading,
+.two-merch-cards.inline-heading,
+.three-merch-cards.inline-heading,
+.four-merch-cards.inline-heading {
+    grid-template-columns: var(--consonant-merch-card-inline-heading-width);
+}
+
+@media screen and ${TABLET_UP} {
+  .two-merch-cards.inline-heading,
+  .three-merch-cards.inline-heading,
+  .four-merch-cards.inline-heading {
+      grid-template-columns: repeat(2, var(--consonant-merch-card-inline-heading-width));
+  }
+}
+
+@media screen and ${DESKTOP_UP} {
+  --consonant-merch-card-inline-heading-width: 378px;
+
+  .three-merch-cards.inline-heading,
+  .four-merch-cards.inline-heading {
+      grid-template-columns: repeat(3, var(--consonant-merch-card-inline-heading-width));
+  }
+}
+
+@media screen and ${LARGE_DESKTOP} {
+  .four-merch-cards.inline-heading {
+      grid-template-columns: repeat(4, var(--consonant-merch-card-inline-heading-width));
+  }
+}
+`;
