+++ libs/features/mas/web-components/src/variants/product.css.js
@@ -1,0 +1,41 @@
+import { TABLET_UP, DESKTOP_UP, LARGE_DESKTOP,} from '../media.js';
+export const CSS = `
+:root {
+  --consonant-merch-card-product-width: 300px;
+}
+
+/* grid style for product */
+.one-merch-card.product,
+.two-merch-cards.product,
+.three-merch-cards.product,
+.four-merch-cards.product {
+    grid-template-columns: var(--consonant-merch-card-product-width);
+}
+
+/* Tablet */
+@media screen and ${TABLET_UP} {
+    .two-merch-cards.product,
+    .three-merch-cards.product,
+    .four-merch-cards.product {
+        grid-template-columns: repeat(2, var(--consonant-merch-card-product-width));
+    }
+}
+
+/* desktop */
+@media screen and ${DESKTOP_UP} {
+  :root {
+    --consonant-merch-card-product-width: 378px;
+  }
+  .three-merch-cards.product,
+  .four-merch-cards.product {
+      grid-template-columns: repeat(3, var(--consonant-merch-card-product-width));
+  }
+}
+
+/* Large desktop */
+@media screen and ${LARGE_DESKTOP} {
+  .four-merch-cards.product {
+      grid-template-columns: repeat(4, var(--consonant-merch-card-product-width));
+  }
+}
+`;
