+++ libs/features/mas/web-components/src/variants/special-offer.css.js
@@ -1,0 +1,49 @@
+import { TABLET_UP, DESKTOP_UP, LARGE_DESKTOP, MOBILE_LANDSCAPE,} from '../media.js';
+export const CSS = `
+:root {
+  --consonant-merch-card-special-offers-width: 378px;
+}
+
+merch-card[variant="special-offers"] span[is="inline-price"][data-template="strikethrough"] {
+  font-size: var(--consonant-merch-card-body-xs-font-size);
+}
+
+/* grid style for special-offers */
+.one-merch-card.special-offers,
+.two-merch-cards.special-offers,
+.three-merch-cards.special-offers,
+.four-merch-cards.special-offers {
+  grid-template-columns: minmax(300px, var(--consonant-merch-card-special-offers-width));
+}
+
+@media screen and ${MOBILE_LANDSCAPE} {
+  :root {
+    --consonant-merch-card-special-offers-width: 302px;
+  }
+} 
+  
+@media screen and ${TABLET_UP} {
+  :root {
+    --consonant-merch-card-special-offers-width: 302px;
+  }
+  .two-merch-cards.special-offers,
+  .three-merch-cards.special-offers,
+  .four-merch-cards.special-offers {
+      grid-template-columns: repeat(2, minmax(300px, var(--consonant-merch-card-special-offers-width)));
+  }
+}
+
+/* desktop */
+@media screen and ${DESKTOP_UP} {
+  .three-merch-cards.special-offers,
+  .four-merch-cards.special-offers {
+    grid-template-columns: repeat(3, minmax(300px, var(--consonant-merch-card-special-offers-width)));
+  }
+}
+
+@media screen and ${LARGE_DESKTOP} {
+  .four-merch-cards.special-offers {
+    grid-template-columns: repeat(4, minmax(300px, var(--consonant-merch-card-special-offers-width)));
+  }
+}
+`;
